Replace GtkStatusIcon with AppIndicator3
