+++ debian/control
@@ -35,6 +35,7 @@ Depends:
     python3-mutagen (>= 1.25),
     xdg-utils
 Recommends:
+    gir1.2-appindicator3-0.1,
     python3-miniupnpc
 Description: graphical client for SoulSeek P2P network (Gtk+ / Python 3)
  Nicotine+ is a client for SoulSeek, a light and efficient file sharing
