+++ pynicotine/gtkgui/chatrooms.py
@@ -1406,7 +1406,7 @@ class ChatRoom:
                     if self.roomsctrl.ChatNotebook.get_current_page() != self.roomsctrl.ChatNotebook.page_num(self.roomsctrl.joinedrooms[self.room].Main) or \
                        self.frame.MainNotebook.get_current_page() != self.frame.MainNotebook.page_num(self.frame.chathbox) or \
                        not self.frame.is_mapped:
-                        if self.room not in self.frame.TrayApp.tray_status["hilites"]["rooms"]:
+                        if self.room not in self.frame.hilites["rooms"]:
                             self.frame.Notifications.Add("rooms", user, self.room, tab=True)
 
             else:
