+++ pynicotine/gtkgui/frame.py
@@ -45,6 +45,7 @@ from pynicotine import slskmessages
 from pynicotine import slskproto
 from pynicotine.gtkgui import imagedata
 from pynicotine.gtkgui import nowplaying
+from pynicotine.gtkgui import tray
 from pynicotine.gtkgui import utils
 from pynicotine.gtkgui.about import AboutDialog
 from pynicotine.gtkgui.about import AboutFiltersDialog
@@ -234,6 +235,13 @@ class NicotineFrame:
         self.bindip = bindip
         self.port = port
         self.got_focus = False
+
+        try:
+            gi.require_version('AppIndicator3', '0.1')
+            from gi.repository import AppIndicator3  # noqa: F401
+            self.appindicator = True
+        except (ImportError, ValueError):
+            self.appindicator = False
 
         try:
             gi.require_version('Notify', '0.7')
@@ -568,7 +576,14 @@ class NicotineFrame:
         self.SetUserStatus(_("Offline"))
 
         self.Notifications = Notifications(self)
-        self.TrayApp = TrayApp(self)
+
+        if self.appindicator:
+            self.TrayApp = tray.TrayApp(self)
+
+        self.hilites = {
+            "rooms": [],
+            "private": []
+        }
 
         self.UpdateBandwidth()
         self.UpdateTransferButtons()
@@ -644,7 +659,7 @@ class NicotineFrame:
         self.UpdateDownloadFilters()
 
         # Create the trayicon if needed
-        if use_trayicon and config["ui"]["trayicon"]:
+        if self.appindicator and use_trayicon and config["ui"]["trayicon"]:
             self.TrayApp.Create()
 
         # Deactivate public shares related menu entries if we don't use them
@@ -725,13 +740,13 @@ class NicotineFrame:
         if not self.np.config.sections["ui"]["exitdialog"]:
             return False
 
-        if self.TrayApp.HAVE_TRAYICON and self.np.config.sections["ui"]["exitdialog"] == 2:
+        if self.appindicator and self.TrayApp.IsTrayIconVisible() and self.np.config.sections["ui"]["exitdialog"] == 2:
             if self.is_mapped:
                 self.MainWindow.unmap()
                 self.is_mapped = False
             return True
 
-        if self.TrayApp.HAVE_TRAYICON:
+        if self.appindicator and self.TrayApp.IsTrayIconVisible():
             option = QuitBox(
                 self,
                 title=_('Close Nicotine+?'),
@@ -1059,7 +1074,7 @@ class NicotineFrame:
             if checkbox:
                 self.np.config.sections["ui"]["exitdialog"] = 0
 
-            if self.TrayApp.trayicon:
+            if self.appindicator and self.TrayApp.trayicon:
                 self.TrayApp.destroy_trayicon()
 
             self.MainWindow.destroy()
@@ -1571,7 +1586,7 @@ class NicotineFrame:
         self.np.config.writeConfig()
 
         # Cleaning up the trayicon
-        if self.TrayApp.trayicon:
+        if self.appindicator and self.TrayApp.trayicon:
             self.TrayApp.destroy_trayicon()
 
         # Closing up all shelves db
@@ -1615,8 +1630,9 @@ class NicotineFrame:
 
     def OnConnect(self, widget):
 
-        self.TrayApp.tray_status["status"] = "trayicon_connect"
-        self.TrayApp.SetImage()
+        if self.appindicator:
+            self.TrayApp.tray_status["status"] = "trayicon_connect"
+            self.TrayApp.SetImage()
 
         if self.np.serverconn is not None:
             return
@@ -1658,8 +1674,9 @@ class NicotineFrame:
 
         self.SetUserStatus(_("Offline"))
 
-        self.TrayApp.tray_status["status"] = "trayicon_disconnect"
-        self.TrayApp.SetImage()
+        if self.appindicator:
+            self.TrayApp.tray_status["status"] = "trayicon_disconnect"
+            self.TrayApp.SetImage()
 
         self.Searches.interval = 0
         self.chatrooms.ConnClose()
@@ -1706,8 +1723,9 @@ class NicotineFrame:
 
         self.SetUserStatus(_("Offline"))
 
-        self.TrayApp.tray_status["status"] = "trayicon_disconnect"
-        self.TrayApp.SetImage()
+        if self.appindicator:
+            self.TrayApp.tray_status["status"] = "trayicon_disconnect"
+            self.TrayApp.SetImage()
 
         self.uploads.ConnClose()
         self.downloads.ConnClose()
@@ -1725,8 +1743,9 @@ class NicotineFrame:
         if self.away == 0:
             self.SetUserStatus(_("Online"))
 
-            self.TrayApp.tray_status["status"] = "trayicon_connect"
-            self.TrayApp.SetImage()
+            if self.appindicator:
+                self.TrayApp.tray_status["status"] = "trayicon_connect"
+                self.TrayApp.SetImage()
 
             autoaway = self.np.config.sections["server"]["autoaway"]
 
@@ -1737,8 +1756,9 @@ class NicotineFrame:
         else:
             self.SetUserStatus(_("Away"))
 
-            self.TrayApp.tray_status["status"] = "trayicon_away"
-            self.TrayApp.SetImage()
+            if self.appindicator:
+                self.TrayApp.tray_status["status"] = "trayicon_away"
+                self.TrayApp.SetImage()
 
         self.SetWidgetOnlineStatus(True)
 
@@ -1815,13 +1835,15 @@ class NicotineFrame:
         if self.away == 0:
             self.SetUserStatus(_("Online"))
 
-            self.TrayApp.tray_status["status"] = "trayicon_connect"
-            self.TrayApp.SetImage()
+            if self.appindicator:
+                self.TrayApp.tray_status["status"] = "trayicon_connect"
+                self.TrayApp.SetImage()
         else:
             self.SetUserStatus(_("Away"))
 
-            self.TrayApp.tray_status["status"] = "trayicon_away"
-            self.TrayApp.SetImage()
+            if self.appindicator:
+                self.TrayApp.tray_status["status"] = "trayicon_away"
+                self.TrayApp.SetImage()
 
         self.np.queue.put(slskmessages.SetStatus(self.away and 1 or 2))
         self.privatechats.UpdateColours()
@@ -1969,7 +1991,8 @@ class NicotineFrame:
         self.DownStatus.push(self.down_context_id, _("Down: %(num)i users, %(speed).1f KB/s") % {'num': usersdown, 'speed': down})
         self.UpStatus.push(self.up_context_id, _("Up: %(num)i users, %(speed).1f KB/s") % {'num': usersup, 'speed': up})
 
-        self.TrayApp.SetToolTip(_("Nicotine+ Transfers: %(speeddown).1f KB/s Down, %(speedup).1f KB/s Up") % {'speeddown': down, 'speedup': up})
+        if self.appindicator:
+            self.TrayApp.SetToolTip(_("Nicotine+ Transfers: %(speeddown).1f KB/s Down, %(speedup).1f KB/s Up") % {'speeddown': down, 'speedup': up})
 
     def BanUser(self, user):
         if self.np.transfers is not None:
@@ -2310,15 +2333,12 @@ class NicotineFrame:
         self.UpdateDownloadFilters()
         self.np.config.writeConfiguration()
 
-        if not config["ui"]["trayicon"] and self.TrayApp.HAVE_TRAYICON:
-            self.TrayApp.destroy_trayicon()
-        elif config["ui"]["trayicon"] and not self.TrayApp.HAVE_TRAYICON:
-            if self.TrayApp.trayicon is None and not self.TrayApp.TRAYICON_CREATED:
+        if self.appindicator:
+            if not config["ui"]["trayicon"] and self.TrayApp.IsTrayIconVisible():
+                self.TrayApp.destroy_trayicon()
+            elif config["ui"]["trayicon"] and not self.TrayApp.IsTrayIconVisible():
                 self.TrayApp.Load()
-            else:
-                self.TrayApp.HAVE_TRAYICON = True
-
-            self.TrayApp.Draw()
+                self.TrayApp.Draw()
 
         if needcompletion:
             self.chatrooms.roomsctrl.UpdateCompletions()
@@ -3385,21 +3405,23 @@ class Notifications:
 
     def Add(self, location, user, room=None, tab=True):
 
-        hilites = self.frame.TrayApp.tray_status["hilites"]
-
         if location == "rooms" and room is not None and user is not None:
-            if room not in hilites["rooms"]:
-                hilites["rooms"].append(room)
+            if room not in self.frame.hilites["rooms"]:
+                self.frame.hilites["rooms"].append(room)
                 self.sound("room_nick", user, place=room)
-                self.frame.TrayApp.SetImage()
+
+                if self.frame.appindicator:
+                    self.frame.TrayApp.SetImage()
         elif location == "private":
-            if user in hilites[location]:
-                hilites[location].remove(user)
-                hilites[location].append(user)
-            elif user not in hilites[location]:
-                hilites[location].append(user)
+            if user in self.frame.hilites[location]:
+                self.frame.hilites[location].remove(user)
+                self.frame.hilites[location].append(user)
+            elif user not in self.frame.hilites[location]:
+                self.frame.hilites[location].append(user)
                 self.sound(location, user)
-                self.frame.TrayApp.SetImage()
+
+                if self.frame.appindicator:
+                    self.frame.TrayApp.SetImage()
 
         if tab and self.frame.np.config.sections["ui"]["urgencyhint"] and not self.frame.got_focus:
             self.frame.MainWindow.set_urgency_hint(True)
@@ -3421,32 +3443,33 @@ class Notifications:
     def Clear(self, location, user=None, room=None):
 
         if location == "rooms" and room is not None:
-            if room in self.frame.TrayApp.tray_status["hilites"]["rooms"]:
-                self.frame.TrayApp.tray_status["hilites"]["rooms"].remove(room)
+            if room in self.frame.hilites["rooms"]:
+                self.frame.hilites["rooms"].remove(room)
             self.SetTitle(room)
         elif location == "private":
-            if user in self.frame.TrayApp.tray_status["hilites"]["private"]:
-                self.frame.TrayApp.tray_status["hilites"]["private"].remove(user)
+            if user in self.frame.hilites["private"]:
+                self.frame.hilites["private"].remove(user)
             self.SetTitle(user)
 
-        self.frame.TrayApp.SetImage()
+        if self.frame.appindicator:
+            self.frame.TrayApp.SetImage()
 
     def SetTitle(self, user=None):
 
-        if self.frame.TrayApp.tray_status["hilites"]["rooms"] == [] and self.frame.TrayApp.tray_status["hilites"]["private"] == []:
+        if self.frame.hilites["rooms"] == [] and self.frame.hilites["private"] == []:
             # Reset Title
             if self.frame.MainWindow.get_title() != _("Nicotine+") + " " + version:
                 self.frame.MainWindow.set_title(_("Nicotine+") + " " + version)
         else:
             # Private Chats have a higher priority
-            if len(self.frame.TrayApp.tray_status["hilites"]["private"]) > 0:
-                user = self.frame.TrayApp.tray_status["hilites"]["private"][-1]
+            if len(self.frame.hilites["private"]) > 0:
+                user = self.frame.hilites["private"][-1]
                 self.frame.MainWindow.set_title(
                     _("Nicotine+") + " " + version + " :: " + _("Private Message from %(user)s") % {'user': user}
                 )
             # Allow for the possibility the username is not available
-            elif len(self.frame.TrayApp.tray_status["hilites"]["rooms"]) > 0:
-                room = self.frame.TrayApp.tray_status["hilites"]["rooms"][-1]
+            elif len(self.frame.hilites["rooms"]) > 0:
+                room = self.frame.hilites["rooms"][-1]
                 if user is None:
                     self.frame.MainWindow.set_title(
                         _("Nicotine+") + " " + version + " :: " + _("You've been mentioned in the %(room)s room") % {'room': room}
@@ -3571,155 +3594,6 @@ class Notifications:
                 self.frame.gstreamer.play(path)
             else:
                 os.system("%s %s &" % (command, path))
-
-
-class TrayApp:
-
-    def __init__(self, frame):
-        self.frame = frame
-        self.trayicon = None
-        self.TRAYICON_CREATED = 0
-        self.HAVE_TRAYICON = False
-        self.tray_status = {
-            "hilites": {
-                "rooms": [],
-                "private": []
-            },
-            "status": "trayicon_disconnect",
-            "last": "trayicon_disconnect"
-        }
-        self.CreateMenu()
-
-    def HideUnhideWindow(self, widget):
-
-        if self.frame.is_mapped:
-            self.frame.MainWindow.unmap()
-            self.frame.is_mapped = False
-        else:
-            self.frame.MainWindow.map()
-            # weird, but this allows us to easily a minimized nicotine from one
-            # desktop to another by clicking on the tray icon
-            if self.frame.minimized:
-                self.frame.MainWindow.present()
-
-            self.frame.MainWindow.grab_focus()
-            self.frame.is_mapped = True
-
-            self.frame.chatrooms.roomsctrl.ClearNotifications()
-            self.frame.privatechats.ClearNotifications()
-
-    def Create(self):
-
-        self.Load()
-        self.Draw()
-
-    def Load(self):
-
-        trayicon = gtk.StatusIcon()
-        self.trayicon = trayicon
-        self.HAVE_TRAYICON = True
-
-    def destroy_trayicon(self):
-
-        if not self.TRAYICON_CREATED:
-            return
-
-        self.TRAYICON_CREATED = 0
-        self.HAVE_TRAYICON = False
-        self.tray_popup_menu.destroy()
-        self.trayicon.set_visible(False)
-        self.trayicon = None
-
-    def Draw(self):
-
-        if not self.HAVE_TRAYICON or self.trayicon is None or self.TRAYICON_CREATED:
-            return
-
-        self.TRAYICON_CREATED = 1
-
-        self.trayicon.set_visible(True)
-        self.trayicon.connect("popup-menu", self.OnStatusIconPopup)
-        self.trayicon.connect("activate", self.OnStatusIconClicked)
-
-        self.SetImage(self.tray_status["status"])
-        self.SetToolTip("Nicotine+")
-
-    def SetImage(self, status=None):
-
-        # Abort if Trayicon module wasn't loaded
-        if not self.HAVE_TRAYICON or self.trayicon is None or not self.TRAYICON_CREATED:
-            return
-
-        try:
-            if status is not None:
-                self.tray_status["status"] = status
-
-            # Check for hilites, and display hilite icon if there is a room or private hilite
-            if self.tray_status["hilites"]["rooms"] == [] and self.tray_status["hilites"]["private"] == []:
-                # If there is no hilite, display the status
-                icon = self.tray_status["status"]
-            else:
-                icon = "trayicon_msg"
-
-            if icon != self.tray_status["last"]:
-                self.tray_status["last"] = icon
-
-            self.trayicon.set_from_pixbuf(self.frame.images[icon])
-
-        except Exception as e:
-            log.addwarning(_("ERROR: cannot set trayicon image: %(error)s") % {'error': e})
-
-    def CreateMenu(self):
-
-        try:
-
-            self.tray_popup_menu_server = popup0 = PopupMenu(self, False)
-
-            popup0.setup(
-                ("#" + _("Connect"), self.frame.OnConnect),
-                ("#" + _("Disconnect"), self.frame.OnDisconnect)
-            )
-
-            self.tray_popup_menu = popup = PopupMenu(self, False)
-
-            popup.setup(
-                ("#" + _("Hide / Show Nicotine+"), self.HideUnhideWindow),
-                (1, _("Server"), self.tray_popup_menu_server, self.OnPopupServer),
-                ("#" + _("Settings"), self.frame.OnSettings),
-                ("#" + _("Send Message"), self.frame.OnOpenPrivateChat),
-                ("#" + _("Lookup a User's IP"), self.frame.OnGetAUsersIP),
-                ("#" + _("Lookup a User's Info"), self.frame.OnGetAUsersInfo),
-                ("#" + _("Lookup a User's Shares"), self.frame.OnGetAUsersShares),
-                ("$" + _("Toggle Away"), self.frame.OnAway),
-                ("#" + _("Quit"), self.frame.OnExit)
-            )
-
-        except Exception as e:
-            log.addwarning(_('ERROR: tray menu, %(error)s') % {'error': e})
-
-    def OnPopupServer(self, widget):
-
-        items = self.tray_popup_menu_server.get_children()
-
-        if self.tray_status["status"] == "trayicon_disconnect":
-            items[0].set_sensitive(True)
-            items[1].set_sensitive(False)
-        else:
-            items[0].set_sensitive(False)
-            items[1].set_sensitive(True)
-        return
-
-    def OnStatusIconClicked(self, status_icon):
-        self.HideUnhideWindow(None)
-
-    def OnStatusIconPopup(self, status_icon, button, activate_time):
-
-        if button == 3:
-            self.tray_popup_menu.popup(None, None, None, None, button, activate_time)
-
-    def SetToolTip(self, string):
-        if self.trayicon is not None:
-            self.trayicon.set_tooltip_text(string)
 
 
 class gstreamer:
