+++ pynicotine/gtkgui/imagedata.py
@@ -7,10 +7,10 @@ notify = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00`\x00\x00\x00`\x08\x06
 offline = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x10\x00\x00\x00\x10\x08\x06\x00\x00\x00\x1f\xf3\xffa\x00\x00\x00\x06bKGD\x00\xfd\x00\xd8\x00\x08\xc4\xa6\xc3)\x00\x00\x00\tpHYs\x00\x00\x0b\x13\x00\x00\x0b\x13\x01\x00\x9a\x9c\x18\x00\x00\x00\x07tIME\x07\xd6\x07\x08\r"/{\x9cIG\x00\x00\x00\x19tEXtComment\x00Created with GIMPW\x81\x0e\x17\x00\x00\x02\xe9IDAT8\xcbu\x93Mh\x1ce\x18\xc7\x7f\xef\xbc3\xb3\xb3\xbb\x93f7\xbb\xb5\xa1\xf6\x0b\xac%\x91\xbaXm \xda\x98\x14B\x03\xa5\x17[\xb5F\xa9`-zT\xf4RT\xbcY<\x89\xc7\x1e*\x94\x80\xa7\x9e\xbch\x1b\xc4\xd4R\xeb\xa5T\\I\x11\xb5\xa4\xa9\xc9\xeeNv\x136\x1f\xeb\xcc\xbc\xb3\xf3\xe1e\xa3\xdb\xa0\xff\xcb\xc3\x03\x0f\xbf\xe7[\xf2\xafL\xe0\xe0\xe5\xa1\xa7\xa7F\n\xc5\xd1o\xebu\x07H\x01.\x10\xf1?\x12\x1d\x9b\x02\x0e\xdf\x19\x1f\xbb\xfad.gK\xdd\x12"\x95\xc1\x8d\xe3\xd0\xfer\xeay\xe0\' \xf8/\x80\x06\x18@\xe9\xc6\xe8\x91\x1fRR\xf6\xa88\x16\x1f\x95\xcb\xbcq\xeb&\x19]\xd7\xef\xbdx\xfak`gW\xb2\x87\x01\x83=v\xe9\xc3\x81\x03SU\xdf\xa7\xe2z,\xba.G\x8by\x9c\xbf6h\xac7\xd9mh\xf9\xe7\x0e\x8b\x11 \xb3\xa5r\x13\x90rb\xc7#g\x1f\xcbfN\xaf\x86m\xb1\x11E\xb8aD\x9fi2\xb9{\x179\t\xe9\x919q\xe6\xfd\x8d\x17\xfeX\x88\xa6\xef\xfe\x9eT\x00\t\xe4?+\x1d<\x95\x95\xd2\x16@\x0e8t\xfe\xc0\xfe\xef\xa4\xa6\x89\xac\x94\x14S&\xfd\x96\xc5\xe0)\x87\xedC+\xa4\xf3\xed\xc4*\xa9g\x81\x85\x93\x8f\xee\x1c8\xd4\xbbmb\xb8\x90\x7f\xc9\xf1U\xb3{\x88c\xb3\'NL;\xad\x16\xa6\x94DQ\xc0p_\x9e\xcf\xefWf\x8b\xa6\xae\x820\xfc\xb3\xdf4\xfb\xf7\xdb\xe9\x01\x95\xc4\xd6\x8a\xaf\xa2\x99\xc6\xf2W\x9b\x00\x1d\x18\n\'\'\x7fL\xa4DD\x11\xc2\xb0H\xa2\x08\x91\x84\x88\xa4\x8d\xa7\x82\xa4\xe6{I\xcd\xf3Y\r\x02\xbeq\x96\xae\\\x9c\x9b?\xafw\x00\xdb\xae\x8e\x1f\xbb\xa0\x01\t\x80aA\xdcF\x0b\xdb\x10\x074\x94\xa2\xea\xf9\xa2\xe6y\xc2Q*\xbe\xbb\xb6\xfe\xdb\xc5\xb9\xf9\x8f\x81\xca&@\xbb\xbf\xb1\xb6\xde,\x14\xe2\x9cD\x13\xca%\x89\x15U\xd7\xa3\xe2+j\x9eG]\x058\xbe\x1f_o,_\xbe\xdeX\xfe\x14x\x00D\x9b\x00}\xb4\x98+eb_\x08/\xa6\x1e\x04,\xba.\x0f\\\x8f\x9a\xafXk\xb7\xe93tl]j5\xdf\x9f\xef\xb4\xac=t\x89G\xb7\x17\x8f_z\xe6\xa9++J\xc9\xba\x1f\x88f;\x10\xaf_\xfb\x95\xef/\xec\xa0<\x93\xc5\x12\x1a\xb6nDY\xc3\x88zM#\x1c\xbfq\xebq\xa0\xda}]\xbdg\xf7\xedyk\xb0\xc7~\xd3\xd0\xb4}\xbb&\x1dy\xf2\xb5\x9693\x9d\xacL\xbc\xad^I[4\n9\xe4\xa2\x83\x07\xf8@\x15\x08\xf4.\x80W|\xb9z{v)\xf1[\xbf\xd8\xc3_\x9c\t^m-%\xc9\x9d\xdb\xb2\x0c\x94\xc7\x86\xb4\xe6\xf1Q\xadG\x08\x82w>\t]\xba\xd6\xf7\x8f\xf6\xee\x15{\xce\x9d\x93\xefe2f\xa67V\xe2\xe7\x85\xf4\xea\x07\x97V\xdf\x05\x9a\xd7n\xc6\xf1\xb1#Z\x94J!;\xad\'ly\x10\x01\x0c\x02O\x00\xaa\x13\x10\x00\xf3\xc0\xbd\x8e\x9fl}\xa6\xbf\x01\x18\x97;(\xac\x8a\xd9\x9b\x00\x00\x00\x00IEND\xaeB`\x82'
 online = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x10\x00\x00\x00\x10\x08\x06\x00\x00\x00\x1f\xf3\xffa\x00\x00\x00\x06bKGD\x00\xfd\x00\xd8\x00\x08\xc4\xa6\xc3)\x00\x00\x00\tpHYs\x00\x00\x0b\x13\x00\x00\x0b\x13\x01\x00\x9a\x9c\x18\x00\x00\x00\x07tIME\x07\xd6\x07\x08\x0c\x150gQ\x8e\xb1\x00\x00\x00\x19tEXtComment\x00Created with GIMPW\x81\x0e\x17\x00\x00\x02\xfbIDAT8\xcbu\x93MlTe\x14\x86\x9f\xef\xfe\xcc\xdcig\xa6\xd3\x81\xc1Nk(T\r\xb6PEM0!\x9641\x12\xc1\x85@U4\x1a""\xa6\x0bW\x1a6\xa2;\xa3\xc1\xc8\xd6\x10\x15!\x06\x16\xe8\xa6.\x0c*\xa6\x1a7Z\x82\xd6P\xdb\x08i`\x06\x9c\x9f\xb6\xc30?\xbds\xbf{\xe7\xfe\xb8\xb9\x8d5\x81\xb3ys6\xcf9\xe7\xcd{T\xfe+\x1d\xe8\x05\xfa\x07\x86\xc7\x9f\xcfn\xdc\xfb\xccRa2\x0f,\x03>w)\x11\xaa\x06\x0c\xee>8s\xa1#\xd9\x9fq\x1d70\x1bU\xa7^\xf9k\xfe\xe2\xf7{\x0e\x023\x80s\'\x80\x12jz\xe4\xd9\x89O\x93k\x1fX7\xfb\xeb\xfb\xca\x8d+_\xab\x9a\xde\x15K\xa4\x07\xb7\x8c\xec\xfbm\n\xd8\x14\x0e\xb9# \xbesT\xd9\x96Ho\x1ev\xed@\xc8V\r-\x92\xa5\xed\xd4\xb1[ea\x9b\x0b\xea\xfd\x0f\x1fy7\xd1\xbd9\xbbjc\x01D\x00U}|\xabx\xf9\x9b\x93\xc9\xb3\xf8=\xc6L\xfe1R\x99m\x08!h5\xf3H\xb3L[.a$\xd6\x0f\x10x\xbf\xd4+\xd3\xd7\xc3\xa1\xdd[\xb6\x1f\xdf\xa7j\x9dqm\xea\xcf`k\xd4\x93\xeaSC\x1f\xa38\xd3|\xf6\xdd\x01lk\x11\xd7m\xe2\xb5M\x04\x8a\xbc<y\xe0E\xe02\x90\xe9\x1d\xd8\xfb`r\xcd#;\x13\xe9\xe1\xe7"F\xe6\xb6\x00\x06\x80{\x80\'^:\xe2|\xb4\xdcXD\x08\r\xd7v\t\x82\x18\xe5\xfc\x17\x7f\xc4:\xfb\xea\xb6U\xafh\xd1u}\x1d\x9d\xf7\rz\x9em8\xf2\x96W)LNh\xc05\xa0\x04t\xf8\xbeB<\xd9\x83\xef\x0b\xb4\x94\xc0\xf7\x03\xba\xd6\xbe\xf5\xa8\xeb\x08l\xcb\n,\xb3\x14H\xb3D\xdb\xa9\xb1\x90\xff\xf6\xab\xdc\xdc\x89\xa3+\xce\xc6\x9e\xdc\x7f\xfe(\x81\x02\x02\xf4(x.\xb8\x8e\xc0\x91`[KH\xb3(\xa4U\x12v\xab\xec7\xaa\xb3Wrs\'\xde\x03\n+\x00\xa5q\xfbz#\x9d\xad\xf9\n)\xc5\xb1\xc0\xb6\x02,\xb3\x844\x0bH\xb3\x88#\x97\x90f\xc9\xaf\x14\x7f>U)\xfe\xf4!\x90\x07\xbc\x15\x80\xd6\x9d\xd9\xf1P\xdb\x8e\t\xdb\x02G.\xd2Z\xfe\x07\xab\x99\xc3n\x95i;5\xb4H\x1aU\x8f+\xb2U\xca\x85\x99P\x00\xd4\x10\xd0\xae\xdd\x9a\xbb\xb6&;:\xb6\\\x9f\x17f\xfd*V3/.\x9e;\xc4b!\xc5\xdf7\xfbPT\r=\xda\xed\xf5\xac\xdf5\xb2a\xe8\x8d\xc37\xaf~y\x12h\x8aU\xa1J\xf6\x0f\x1e~-\x9e\xda\xf4\xba\xa2h\x1b^\x19;\'\xde\x1e\x9b\xee\xb8\xf0cda\xd7\xa1\xc6\xfe\x98A5\x93F\xdc(b\x0324\xfe\x7f\xf1V\xcf\x9c\xd1\x87\x8f\x1d\xd3wo\x1f\xdd\xf1A\xb5\x98p\xaa\x97\x0c\xff\xcdW{\x7f\x07\xee}zDQ\xcf\x1e\xd7\x93\xe7?\xd7\x8d\xbb=V\xf4\xf4i\xfd\x85j\xd5\x98*\x14\x8cY7\x1f\xf5\x7f8eH`\x1cH\x01\xca;\xe3j\xd7\xc4\'z\xe7\xaaH\xb3\xfa\x84\x18\xb0\x11\x18\x02l \x08W\xcc\x01\xf3a\xaf\x87\xea\x03\x1e\xc0\xbfm\x08G\x0c\x84\x99\xf6]\x00\x00\x00\x00IEND\xaeB`\x82'
 plugin = b"\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x10\x00\x00\x00\x10\x08\x06\x00\x00\x00\x1f\xf3\xffa\x00\x00\x00\x04gAMA\x00\x00\xaf\xc87\x05\x8a\xe9\x00\x00\x00\x19tEXtSoftware\x00Adobe ImageReadyq\xc9e<\x00\x00\x01\xe1IDAT8\xcb\xa5\x92=H\x95q\x14\xc6\x7f\xaff\x05~\\?\x91\xc4,\x88\x9byoX\xa3{KC\xb5\xb47\x1a\xd1Pc\x05M\x15\xb45\xba4EkS\x14$\xa48V\x16*w\xb2\xb8\x82\x8aJ\xddB%\xcb\xee{\x9es\x1a^1\xe3\xaa\x8bg9\xf0\xe7\x9c\xe7\x7f\x9e\xdf9IDp\x908\xb4\xdb\xe3\x83\xa9\xbb%w\x15%\x0fI\xaf\x1f\r=\xb9\xb4\x97@]m\xf3\x9dv\x9cb\x7fS?\xf9\x96|b\xd2\xd0~\x13\xd4\x08\xdc?\xff\xf8G\xd5\xaa\x1bi\xa4lT70\xb3\x99\xfd\x04\x92\x88\xe0\xe1\xf4\xbdp\x17\x92\xa3,\xc7@\xdb\x99\xc4\xdc\x98^\x9eY3\x93$\xd5\x99\x19O\xaf<o\xaba \x17\xa7\x1b\xf3x8\x1eA\xe0\x89\xdc\x89\x80BW!\xa7\x10\x11\xce\xfb\xf2\xc7\xdd!JB\xe1x8K?\xe7Y\xdd\\\xc5\xc2\xf0\x08\xea\x92z\xdc\x83S\xb9<f\xb6\x87\x80\x89R\xa5DKC\x8e\xee\xc6\x1e\xbe\xfd\xaa\xa0Pf'\xfe \x17\x86\xe8h\xea\xe0\xea\xb3\xcbaf\x98\x89W\xc3\xa3I\xb2\xf3\x0en\x8f]\x8fBW\x81\xd2\xf7)\xe4\x8e\x87\xb0p\x14\xa2\xa5!Gos\x1f\xeeA\xe0\x8cMN\xf0\xe6\xc6X\xf2\xdf\x1d\x98\t\x850eY[\xcdra\xf5B\x9eA\x8e\xad\xdam\x0b\xb7\xde\x0e\x87\x99h=\xda\x8a\xdcI#\xcd\x98\xb80\xcf\xd8\xf46\xf71;\xff\x99\x85\x95\xc5\xaa\xa4\xc3\xd2\x0e\x013Q\xec\x1e@\x04s\xebe\n\xed\x83\xdb\x1b\xf1p\xcak_\x90;\x0b+\x8b\x8c\xde\x1c?R\x0315\xc3\x10S_?q\xb6\xf3\x1c\xe5\x959\x96*\xcbH\xa23\xd7\xc9\xc9\x9e\x13\x99\x8d\xbd\xb6`\xa9!w\x8a\x1d\x83\x98\x9c\xa5\xca2/\xae\xbdL\x00.\x8e\\\x88\xbec\xc7Ie\xd8\xd6\xd8\xb5\x02f|\x98\x9d\xc4\x94\xfdb\xa6\xcd\x7f`\xcd\xc7\xdfM \xf9o\xb9\xd6w=\xe5\x83\xc4_\xc7\xc4m\xd06E\xe5X\x00\x00\x00\x00IEND\xaeB`\x82"
-trayicon_away = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x10\x00\x00\x00\x10\x08\x03\x00\x00\x00(-\x0fS\x00\x00\x00\x04gAMA\x00\x00\xb1\x8f\x0b\xfca\x05\x00\x00\x00 cHRM\x00\x00z&\x00\x00\x80\x84\x00\x00\xfa\x00\x00\x00\x80\xe8\x00\x00u0\x00\x00\xea`\x00\x00:\x98\x00\x00\x17p\x9c\xbaQ<\x00\x00\x024PLTE~\xa3\xff\x00\x00\x00\xff\xc1\x00\xff\xba\x00\x00\x00\xff\xff\xb1\x00\xff\xb0\x00\xd3\xaah\xf6\xad\x0e\xc6\xa9\x88\xa5\xa6\xdb\x9d\xa6\xf1\xa4\xa6\xe0\x91\xa5\xf5\x89\xa4\xff\x85\xa4\xffe\xa1\xff\xec\x00\x00\xf2CC\xedts\xe9\x8d\x8c\xe9\x84\x82\xefZY\xf1**\xc6\xa9\x89\xd6\xab`\xd7\xab^\xd6\xab_\xd0\xaap\xbc\xa8\xa0h\xa1\xff\xf0BA\xe2\x8a\x88\xd7\xac\xa8\xebed\xf2\x10\x11\xd4\xaae\xdf\xabI\xc1\xa9\x96\xeapn\xd9\x98\x95\xefCB\xd4\xaad\xeb\xac+\xea\xac-\xdb\xabT\xe1\x85\x82\xedSR\xca\xaa\x7f\xc1\xa9\x95\xdd\xabN\xe9\xac/\xe0~{\xd0\xa6\xa2\xecNM\xc8\xa9\x84\xd3\xaah\xe1k_\xd5\xaa\xa6\xde~{\xf0--\xc8\xa9\x83\xd3\xaag\xd1\x8fo\xe6a_\xd8\x97\x94\xd0\xaf\xab\xd2\xa7\xa3\xdd\x7f{\xecDC\xff\x00\x00\xd4\xaae\xc9\xa9\x81\xd4-=\xdc~\\\xe7{E\xe4\x7fI\xd3{m\xfd\x00\x00\xc6\xa9\x86S\x9f\xff\xd6\xab^\xc3\xa9\x8cq\xa2\xf9\xd7\xab]\xd3\xaah\xdf\xabI\xd6\xab`\xd4\xaad\xc7\xa9\x86\xd8\xab[\xd1\xaal\xc6\xa9\x88\xd6\xaaa\xc6\xa9\x88\xbf\xa9\x9b\xc8\xa9\x84\xca\xa9~\xc8\xa9\x84\xda\xabV\xcd\xaav\xca\xaa~\xc1\xa9\x95\xd4\xaad\xc8\xaa\x84\xbd\xa8\x9f\xb8\xa8\xabW\xa0\xff\xd7\xab^\xc6\xa9\x89\xc8\xa9\x84\xda\xabW\xd6\xab_\xc0\xa9\x98\xbf\xa9\x9a\xcb\xaaz\xc9\xaa\x7f\xbb\xa8\xa5B\x9d\xff\xca\xa9\x7f\xda\xabV\xd6\xaba\xbf\xa9\x9b\xa2\xa6\xe2\x9e\xa6\xec\x9f\xa6\xeb\xa4\xa6\xdd\xab\xa7\xcc\xaa\xa7\xcd\xab\xa7\xcd\xac\xa7\xcb\xa2\xa6\xe4\x9e\xa6\xed\xa5\xa6\xdc\xd0\xb3\xb1\xd9\xbd\xbb\xd2\xca\xc7\xf0\xad\x1d\xf2\xad\x1a\xf2\xad\x19\xef\xad \xc7\xb8\xb5\xa1\x8b\x89\xd4\x86\x85\xe7\xdd\xd9\xcd\xd2\xcd\xf1\xad\x1c\xf1\xad\x1b\xd3\xd3\xd0\xc7\xc7\xc5ymj\x87\x86\x83\xab\xaf\xab\xce\xae\xaa\xec\xad\'\xf3\xad\x18\xd5\xd0\xcd\xf1\xf2\xf1\xc3\xc5\xc2\xc1\xc3\xc0\xc8\xcc\xc8\xec\xac)\xec\xac(\xdc\xdf\xdd\xf1\xf3\xf2\xe8\xeb\xe9\xce\xc7\xc3\xf3\xad\x17\xeb\xac)\xef\xad"\xf2\xad\x18\xee\xad$\xf1\xad\x1a\xec\xac\'\xe7\xac4\xee\xad"\xed\xad%\xf0\xad\x1f\xee\xad#\xd8\xab[\xd4\xaad\xdd\xabP\xdb\xabU\xeb\xac*\xc9\xa9\x81\xc7\xa9\x86\xeb\xac+\xff\xff\xff6uUa\x00\x00\x00\x86tRNS\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x05\t\x08\x02\x03M\xc0\xf5\xe3\x8c\x1a\x88\xba\xb8\xb7\x99<\x01M\xde\xfb\x9a\x0e\xc6\xe6R\xc0\xedS\xc5\xfe\xfe\xd3\xf5\x85\x9ea\xd9\xfd\xec\xfask\xa9\xbf\xfa\xd2-l\xa8\x7f\x9a\xf1\xfe\xfb\xd5Q\x01\xc4p\x12\xca\xfb\xf8z\x01r\t\xbdh\x03\xd2\xb1\xea\xcf\xaf\x91\xde\xc3k\xdd\xc1h\xa6\xb3}\xdc\xafzd\xc6l\xa4k\x04\xbce\x95\xca\xc8O}\xcc\xc9Q\x02\x8e\xcc\xc8J\n\x11\x10\x05\t\x11\x10\x05\n\x11\x04\x04\x9e}i\x00\x00\x00\x01bKGD\xbb\xd4\xb6wL\x00\x00\x00\tpHYs\x00\x00\x04\xb0\x00\x00\x04\xb0\x01\xe7,\'\xa6\x00\x00\x00\x07tIME\x07\xe0\x08\x1e\x01\x15\n\x8c\x84\xbfr\x00\x00\x01\x05IDAT\x18\xd3c\xe0\xe5\xe3\xe3\xe3\x17``\x14\x14\x12\x16\x11\x15\x13gd\x90\x90\x94\x92\x92\x96\x91\x95\x93Whk\xefPTRfP\xe9\xec\xea\xee\xeeQUS\xef\xed\xeb\x9f0QC\x93Ak\xd2dm\x9d\xcen]\xbd)S\xa7M\x9f1S\x9fAk\xf2,\x03C\xa3\xd9\xc6&s\xe6\xce\x9b\xbf\xc0\xd4\x0c(\xb0\xd0\x9c\xc9b\xf2"K\xab\xc5K\x96.\xb3\xb6\x01\t\xd82\xdbMZd\xef\xe0\xe8\xe4\xec\xe2\xea\xc6\xe0\x0e\x15\xf0\xf0\xf4\xf2\xf6\xf1\xf5c\x84\x0b\xf8\x07\x04._\x11\xc4\x12\x0c\x13X\x19\x12\x1a\xb6juxD$\x83\xfb\x9aE \x81\xee\xae\xc9\xdd\xdd@"\x8a!z\xed\xba\x18f\xbb\xc9\xab\xd6o\xd8\xd8\xdd\xb5iC,C\xdc\xe6-\xf1\xac\t[\xb7%&%\xaf\x9a\x95\x92\x9a\xc6\x90>i{\x06[\xe6\x8e\x9dY\xd99\xdd\xbbr\xd9\x19\x19\xf2\xf2\x0b\n9\x8a\x8aKJ\xcb\xca+*\xab8\x19\x19\xaakj\xeb\xb8\xea\x1b\x1a\x9b\xb8\x9b[j[y\x18\x01\xb2\x8f_\x87\xab1\xc7-\x00\x00\x00%tEXtdate:create\x002016-08-30T01:21:10+02:00^(\x80!\x00\x00\x00%tEXtdate:modify\x002016-08-30T01:21:10+02:00/u8\x9d\x00\x00\x00\x19tEXtSoftware\x00www.inkscape.org\x9b\xee<\x1a\x00\x00\x00\x00IEND\xaeB`\x82'
-trayicon_connect = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x10\x00\x00\x00\x10\x08\x03\x00\x00\x00(-\x0fS\x00\x00\x00\x04gAMA\x00\x00\xb1\x8f\x0b\xfca\x05\x00\x00\x00 cHRM\x00\x00z&\x00\x00\x80\x84\x00\x00\xfa\x00\x00\x00\x80\xe8\x00\x00u0\x00\x00\xea`\x00\x00:\x98\x00\x00\x17p\x9c\xbaQ<\x00\x00\x02(PLTE~\xa3\xff\x00\x00\x00\xff\xc1\x00\xff\xba\x00\x00\x00\xff\xff\xb1\x00\xff\xb0\x00\xd3\xaah\xf6\xad\x0e\xc6\xa9\x88\xa5\xa6\xdb\x9d\xa6\xf1\xa4\xa6\xe0\x91\xa5\xf5\x89\xa4\xff\x85\xa4\xffe\xa1\xff1c\x003f\x042e\x032d\x033e\x035a\x00\xc6\xa9\x89\xd6\xab`\xd7\xab^\xd6\xab_\xd0\xaap\xbc\xa8\xa0h\xa1\xff.]\x005g\x08\x88\xafaW\x98\x1aL\x97\x06M\x96\x06\xd4\xaae\xdf\xabI\xc1\xa9\x963e\x03\x81\xa9XQ\x9a\x0c\xd4\xaad\xeb\xac+\xea\xac-\xdb\xabT@~\x06C\x86\x05\xca\xaa\x7f\xc1\xa9\x95\xdd\xabN\xe9\xac/L\x94\t3e\x03\xc8\xa9\x84\xd3\xaahN\x7f\x12\x93\xbak>y\x063d\x03\xc8\xa9\x83\xd3\xaag\xae\x9cj6k\x05\x8c\xafj;s\x052b\x036^\x00\xd4\xaae\xc9\xa9\x814f\x15Wt\x1c=g\x06=h\x07Gm\x17/^\x00\xc6\xa9\x86S\x9f\xff\xd6\xab^\xc3\xa9\x8cq\xa2\xf9\xd7\xab]\xd3\xaah\xdf\xabI\xd6\xab`\xd4\xaad\xc7\xa9\x86\xd8\xab[\xd1\xaal\xc6\xa9\x88\xd6\xaaa\xc6\xa9\x88\xbf\xa9\x9b\xc8\xa9\x84\xca\xa9~\xc8\xa9\x84\xda\xabV\xcd\xaav\xca\xaa~\xc1\xa9\x95\xd4\xaad\xc8\xaa\x84\xbd\xa8\x9f\xb8\xa8\xabW\xa0\xff\xd7\xab^\xc6\xa9\x89\xc8\xa9\x84\xda\xabW\xd6\xab_\xc0\xa9\x98\xbf\xa9\x9a\xcb\xaaz\xc9\xaa\x7f\xbb\xa8\xa5B\x9d\xff\xca\xa9\x7f\xda\xabV\xd6\xaba\xbf\xa9\x9b\xa2\xa6\xe2\x9e\xa6\xec\x9f\xa6\xeb\xa4\xa6\xdd\xab\xa7\xcc\xaa\xa7\xcd\xab\xa7\xcd\xac\xa7\xcb\xa2\xa6\xe4\x9e\xa6\xed\xa5\xa6\xdc\xa8\xcd\x82\x95\xc3h\xf0\xad\x1d\xf2\xad\x1a\xf2\xad\x19\xef\xad \x96\xc5he\xb1\x1a\\\xa9\x12|\xb4F\xf5\xf9\xf1\xf1\xad\x1c\xf1\xad\x1bT\x9d\x0eT\x9e\rY\xa7\r`\xa4\x1e\xee\xf5\xe8\xa7\xcd\x83\xec\xad\'\xf3\xad\x18\xff\xff\xff\x90\xbfbR\x9c\x0b\xd1\xe5\xbe\xd2\xe5\xbfO\x95\x0b\xec\xac)\xec\xac(\xfb\xfd\xfa\xe7\xf1\xde`\x9f"w\xa5J\xf3\xad\x17\xeb\xac)\xef\xad"\xf2\xad\x18\xee\xad$\xf1\xad\x1a\xec\xac\'\xe7\xac4\xee\xad"\xed\xad%\xf0\xad\x1f\xee\xad#\xd8\xab[\xd4\xaad\xdd\xabP\xdb\xabU\xeb\xac*\xc9\xa9\x81\xc7\xa9\x86\xeb\xac+\x95\xd0\x90\x00\x00\x00\x00\x83tRNS\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x05\t\x08\x02\x1f\xb4\xf1\xf0\xb3\x1d\x88\xba\xb8\xb7\x99<\x01\x16\xed\xfc\xfe\xfe\x85\xc6\xe6R\xab\xfc\xf8\xc5\xfe\xfe\xd3\xf8\xf8\x9ea\xd9\xfd\xfe\xeek\xa9\xdf\xfe\xfa\xa6l\xa8|\xee\xfd\xfa\xec\x13\xc4p \xe7\xfe\xfe\xcc\x1br\t\xbdh\x03\xd2\xb1\xea\xcf\xaf\x91\xde\xc3k\xdd\xc1h\xa6\xb3}\xdc\xafzd\xc6l\xa4k\x04\xbce\x95\xca\xc8O}\xcc\xc9Q\x02\x8e\xcc\xc8J\n\x11\x10\x05\t\x11\x10\x05\n\x11\x04]z~b\x00\x00\x00\x01bKGD\x98v\xd1\x06>\x00\x00\x00\tpHYs\x00\x00\x04\xb0\x00\x00\x04\xb0\x01\xe7,\'\xa6\x00\x00\x00\x07tIME\x07\xe0\x08\x1e\x01\x15\n\x8c\x84\xbfr\x00\x00\x01\x05IDAT\x18\xd3c\xe0\xe5\xe3\xe3\xe3\x17``d\x14\x14\x12\x16\x11\x15cd\x10\x97\x90\x94\x94\x92\x96\x91\x95\x93WhnQTRfPimko\xefPUS\xd7\xe8\xec\xea\xee\xe9\xd5d\xd0\xea\xeb\xd7\xd6im\xd7\xd5\x9b0q\xd2\xe4)S\xf5\x19\xb4\xfa\xa7\x19\x18\x1aM76\x991s\xd6\xec9sM\x81\x02\xf3\xcc\x98\xcc\xfb\xe7[X.X8e\x91\x955H\xc0\x86\xd9\xb6o\xbe\x9d\xbd\xc3\x8c\xc5\x8eN\xce\x0c.P\x01W7w\x0fO/oF\xb8\x80\x8f\xaf\xdf\x92\xa5\xfe,\x010\x81e\x81A\xc1\xcbW\x84\x84\x861\xb8\xac\x9c\x0f\x12ho\xeboo\x07\x12\xe1\x0c\x11\xabVG2\xdb\xf6/_\xb3v]{\xdb\xfa\xb5Q\x0c\xd1\x1b6\xc6\xb0\xc6n\xda\x1c\x17\x9f\xb0|ZbR2CJ\xdf\x96T\xb6\xb4\xad\xdb\xd232\xdb\xb7g\xb132d\xe7\xe4\xe6q\xe4\x17\x14\x16\x15\x97\x94\x96\x95s22TTVUs\xd5\xd4\xd6\xd5s74V5\xf10\x02\x00\xe0O]\xde\xc1\xc8\x9fK\x00\x00\x00%tEXtdate:create\x002016-08-30T01:21:10+02:00^(\x80!\x00\x00\x00%tEXtdate:modify\x002016-08-30T01:21:10+02:00/u8\x9d\x00\x00\x00\x19tEXtSoftware\x00www.inkscape.org\x9b\xee<\x1a\x00\x00\x00\x00IEND\xaeB`\x82'
-trayicon_disconnect = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x10\x00\x00\x00\x10\x08\x03\x00\x00\x00(-\x0fS\x00\x00\x00\x04gAMA\x00\x00\xb1\x8f\x0b\xfca\x05\x00\x00\x00 cHRM\x00\x00z&\x00\x00\x80\x84\x00\x00\xfa\x00\x00\x00\x80\xe8\x00\x00u0\x00\x00\xea`\x00\x00:\x98\x00\x00\x17p\x9c\xbaQ<\x00\x00\x02(PLTE~\xa3\xff\x00\x00\x00\xff\xc1\x00\xff\xba\x00\x00\x00\xff\xff\xb1\x00\xff\xb0\x00\xd3\xaah\xf6\xad\x0e\xc6\xa9\x88\xa5\xa6\xdb\x9d\xa6\xf1\xa4\xa6\xe0\x91\xa5\xf5\x89\xa4\xff\x85\xa4\xffe\xa1\xffEKD\x9b\x9e\x99\xb3\xa9\xa5\xaa\xa7\xa3\x87\x8c\x87\xc6\xa9\x89\xd6\xab`\xd7\xab^\xd6\xab_\xd0\xaap\xbc\xa8\xa0h\xa1\xff\xbb\xa0\x9d\xc3|{\xac\xa8\xa4\x8a\x8e\x89\xd4\xaae\xdf\xabI\xc1\xa9\x96\xb3\xa9\xa5\x9c\xa3\x9f\xd4\xaad\xeb\xac+\xea\xac-\xdb\xabT\xa1\xa9\xa5\xca\xaa\x7f\xc1\xa9\x95\xdd\xabN\xe9\xac/\xa0\xa8\xa3\xc8\xa9\x84\xd3\xaah\xad\xa7\x9f\xc2{y\xba\xa1\x9e\x96\x9b\x96\xc8\xa9\x83\xd3\xaag\xb8\xa1\x83\xc1\x8a\x88\xba\xa1\x9e\x9a\x9f\x9aUTM\xd4\xaae\xc9\xa9\x81\x84\x8f\x91\xb1\xa6\x8c\xbe\xaaw\xbb\xa9|\xa8\xa1\x95cd^\xc6\xa9\x86S\x9f\xff\xd6\xab^\xc3\xa9\x8cq\xa2\xf9\xd7\xab]\xd3\xaah\xdf\xabI\xd6\xab`\xd4\xaad\xc7\xa9\x86\xd8\xab[\xd1\xaal\xc6\xa9\x88\xd6\xaaa\xc6\xa9\x88\xbf\xa9\x9b\xc8\xa9\x84\xca\xa9~\xc8\xa9\x84\xda\xabV\xcd\xaav\xca\xaa~\xc1\xa9\x95\xd4\xaad\xc8\xaa\x84\xbd\xa8\x9f\xb8\xa8\xabW\xa0\xff\xd7\xab^\xc6\xa9\x89\xc8\xa9\x84\xda\xabW\xd6\xab_\xc0\xa9\x98\xbf\xa9\x9a\xcb\xaaz\xc9\xaa\x7f\xbb\xa8\xa5B\x9d\xff\xca\xa9\x7f\xda\xabV\xd6\xaba\xbf\xa9\x9b\xa2\xa6\xe2\x9e\xa6\xec\x9f\xa6\xeb\xa4\xa6\xdd\xab\xa7\xcc\xaa\xa7\xcd\xab\xa7\xcd\xac\xa7\xcb\xa2\xa6\xe4\x9e\xa6\xed\xa5\xa6\xdc\xb7\xa4\xa1\xb7\xa6\xa3\xcaON\xcf++\xca--\xf0\xad\x1d\xf2\xad\x1a\xf2\xad\x19\xef\xad \xc8ML\xdf\x1a\x1a\xe6\x1c\x1d\xdd\x13\x13\xc9\x18\x18\xc1\x89\x87\xf1\xad\x1c\xf1\xad\x1b\xb7\xa5\xa2\xca\'&\xe0\x15\x15\xde\x15\x15\xd9\x0f\x0f\xcd\x03\x03\xc2ml\xec\xad\'\xf3\xad\x18\xc5(\'\xd5\t\t\xd5\x0b\x0b\xd2\x06\x06\xc6\x00\x00\xc2sr\xec\xac)\xec\xac(\xc3\x12\x12\xc8\x00\x00\xc5\x00\x00\xc1??\xc1kj\xf3\xad\x17\xeb\xac)\xef\xad"\xf2\xad\x18\xee\xad$\xf1\xad\x1a\xec\xac\'\xe7\xac4\xee\xad"\xed\xad%\xf0\xad\x1f\xee\xad#\xd8\xab[\xd4\xaad\xdd\xabP\xdb\xabU\xeb\xac*\xc9\xa9\x81\xc7\xa9\x86\xeb\xac+\xff\xff\xff\xd6\xb0\xc0g\x00\x00\x00|tRNS\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x05\t\x08\x02\x03j\xf4\xc0"\x88\xba\xb8\xb7\x99<\x01\xe7\xfe\xae"\xc6\xe6R\xf3\x83\xc5\xfe\xfe\xd3\x8f\x9ea\xd9\xfd\x8fk\xa9\xdc\xfd\xe8Xl\xa8\x83\xfe\xeb^\x03\xc4p%\xd9\xfc\xfa\x95\x04r\t\xbdh\x03\xd2\xb1\xea\xcf\xaf\x91\xde\xc3k\xdd\xc1h\xa6\xb3}\xdc\xafzd\xc6l\xa4k\x04\xbce\x95\xca\xc8O}\xcc\xc9Q\x02\x8e\xcc\xc8J\n\x11\x10\x05\t\x11\x10\x05\n\x11\x04\x9e\x93b\x0c\x00\x00\x00\x01bKGD\xb7\xdd\x00;g\x00\x00\x00\tpHYs\x00\x00\x04\xb0\x00\x00\x04\xb0\x01\xe7,\'\xa6\x00\x00\x00\x07tIME\x07\xe0\x08\x1e\x01\x15\n\x8c\x84\xbfr\x00\x00\x01\x04IDAT\x18\xd3c\xe0\xe5\xe3\xe3\xe3\x17``\x14\x14\x12\xae\xa9\x15\x11ed\x10\x13\x97\x90\x90\x94\x92\x96\x11\x92\xad\xabo\x90\x93W`PlljnnQRVimk\xef\xe8\xecReP\xeb\xeeQ\xd7hl\xd6\xec\xed\xeb\x9f0q\xd2d-\x06\xb5\x9e)\xda:\xbaS\xf5j\xa7M\x9f1s\xd6l}\xa0\xc0\x1c\x03&\xc3\x9e\xb9F\xc6\xf3\xe6/Xhb\n\x120c6\xef\x9ek!o\xb9h\xb6\x95\xb5\r\x83-T\xc0\xce\xde\xc1\xd1\xc9\xd9\x85\x11.\xe0\xea\xe6\xbex\x89\x07\x8b\'L`\xa9\x97\xb7\xcf\xb2\xe5\xbe~\xfe\x0c\xb6+\xe6\x82\x04\x9a\x9bz\x9a\x9b\x81D\x00C\xe0\xcaUA\xcc\xe6=\xcbV\xafY\xdb\xdc\xb4nM0C\xc8\xfa\r\xa1\xaca\x1b7\x85GD.\x9b\x12\x15\x1d\xc3\x10\xdb\xbd9\x8e-~\xcb\xd6\x84\xc4\xa4\xe6m\xc9\xec\x8c\x0c)\xa9i\xe9\x1c\x19\x99Y\xd99\xb9y\xf9\x05\x9c\x8c\x0c\x85E\xc5%\\\xa5e\xe5\x15\xdc\x95U\xc5\xd5<\x8c\x00K\x82\\o\xe3\x00+\xc1\x00\x00\x00%tEXtdate:create\x002016-08-30T01:21:10+02:00^(\x80!\x00\x00\x00%tEXtdate:modify\x002016-08-30T01:21:10+02:00/u8\x9d\x00\x00\x00\x19tEXtSoftware\x00www.inkscape.org\x9b\xee<\x1a\x00\x00\x00\x00IEND\xaeB`\x82'
-trayicon_msg = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x10\x00\x00\x00\x10\x08\x03\x00\x00\x00(-\x0fS\x00\x00\x00\x04gAMA\x00\x00\xb1\x8f\x0b\xfca\x05\x00\x00\x00 cHRM\x00\x00z&\x00\x00\x80\x84\x00\x00\xfa\x00\x00\x00\x80\xe8\x00\x00u0\x00\x00\xea`\x00\x00:\x98\x00\x00\x17p\x9c\xbaQ<\x00\x00\x01\xf2PLTE~\xa3\xff\x00\x00\x00\xff\xc1\x00\xff\xba\x00\xff\xb1\x00\xff\xb0\x00\xd3\xaah\xf6\xad\x0e\xc6\xa9\x88\xa5\xa6\xdb\x9d\xa6\xf1\xa4\xa6\xe0\x91\xa5\xf5\x89\xa4\xff\x85\xa4\xffe\xa1\xff\x00\x00\xff\x00\x00\xff\xc6\xa9\x89\xd6\xab`\xd7\xab^\xd6\xab_\xd0\xaap\xbc\xa8\xa0h\xa1\xff\x00\x00\xff\x00\x00\xff\xd4\xaae\xdf\xabI\xa2\x8d\xa7\x00\x00\xff\x00\x00\xff\x00\x00\xff\x00\x00\xff\x00\x00\xff\x00\x00\xff\xd4\xaad\xeb\xac+\xea\xac-x]\xa2\x08\x07\xfc\x00\x00\xff\x00\x00\xff\x00\x00\xff\x00\x00\xff\x00\x00\xff\x00\x00\xff\xca\xaa\x7f\xc1\xa9\x95\xdd\xabN\xd1\x9aD)$\xe7\x00\x00\xff\x00\x00\xff\x00\x00\xff\xc8\xa9\x84\xd3\xaah\xb5\x98\x8e\x00\x00\xff\x00\x00\xff\x00\x00\xff\xc8\xa9\x83\xd3\xaag\xba\x9c\x8b\x00\x00\xff\x00\x00\xff\x05\x05\xfd\x00\x00\xff\x05\x04\xfd\x00\x00\xff\xd4\xaae\x8at\xa9\x00\x01\xffUF\xc4\xb6\x89bnT\xa5\x15\x12\xf3\x00\x00\xff\x96\x80\xa3\t\x11\xff\xd1\xa7arc\xbc\x00\x00\xffq\xa2\xf9\xd7\xab]\xd3\xaah\xdf\xabI\xd6\xab`\xd4\xaad\xc7\xa9\x86\xd8\xab[\xd1\xaal\xc6\xa9\x88\xd6\xaaa\xc6\xa9\x88\xbf\xa9\x9b\xc8\xa9\x84\xca\xa9~\xc8\xa9\x84\xda\xabV\xcd\xaav\xca\xaa~\xc1\xa9\x95\xd4\xaad\xc8\xaa\x84\xbd\xa8\x9f\xb8\xa8\xabW\xa0\xff\xd7\xab^\xc6\xa9\x89\xc8\xa9\x84\xda\xabW\xd6\xab_\xc0\xa9\x98\xbf\xa9\x9a\xcb\xaaz\xc9\xaa\x7f\xbb\xa8\xa5B\x9d\xff\xca\xa9\x7f\xda\xabV\xd6\xaba\xbf\xa9\x9b\xa2\xa6\xe2\x9e\xa6\xec\x9f\xa6\xeb\xa4\xa6\xdd\xab\xa7\xcc\xaa\xa7\xcd\xab\xa7\xcd\xac\xa7\xcb\xa2\xa6\xe4\x9e\xa6\xed\xa5\xa6\xdc\xf0\xad\x1d\xf2\xad\x1a\xf2\xad\x19\xef\xad \x00\x00\xff\xf1\xad\x1c\xf1\xad\x1b\xec\xad\'\xf3\xad\x18\xec\xac)\xec\xac(\xf3\xad\x17\xe7\xa9,\xef\xad"\xf2\xad\x18\xee\xad$\xf1\xad\x1a\xec\xac\'\xe7\xac4\xee\xad"\xed\xad%\xf0\xad\x1f\xee\xad#\xd8\xab[\xd4\xaad\xdd\xabP\xdb\xabU\xeb\xac*\xc9\xa9\x81\xc7\xa9\x86\xeb\xac+\xff\xff\xff*~w\xc7\x00\x00\x00\x86tRNS\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x05\t\x08\x02\x11d\x88\xba\xb8\xb7\x99<\x01:\xd3\xc6\xe6\\\x15\x1d\x97\x1e\x16\x10\xc5\xfe\xfe\xe5\xc4\xd2\xf2\xf3\xd4\xbdi\x9ea\xd9\xfd\xb9\xf7\x92\x1ak\xa9w\xb3\xb5\x0cl\xa8u\xba\xfe\xe9\xfe\xc2\x0b\xc4\x88\xbf\xd7\xf9\xf9\xd7.\x84@\xbe\x8a!\x03\xd2\xb1\xea\xcf\xaf\x91\xde\xc3k\xdd\xc1h\xa6\xb3}\xdc\xafzd\xc6l\xa4k\x04\xbce\x95\xca\xc8O}\xcc\xc9Q\x02\x8e\xcc\xc8J\n\x11\x10\x05\t\x11\x10\x05\n\x11\x04\x08\xf4J/\x00\x00\x00\x01bKGD\xa5.\xb9J/\x00\x00\x00\tpHYs\x00\x00\x04\xb0\x00\x00\x04\xb0\x01\xe7,\'\xa6\x00\x00\x00\x07tIME\x07\xe0\x08\x1e\x01\x15\n\x8c\x84\xbfr\x00\x00\x00\xfcIDAT\x18\xd3c\xe0\xe1\xe5\xe5\xe5\xe3g`dd\x14\x10\x14\x00\x92\x0cB\xc2""\xa2b\xe2\x12\x8c\x8c\x92R\x92 \x01\xe9\xb6\xf6\x8e\x8eN\x19Y9y\x85.\x05E%e\x06\x95\xee\x1eU\xb5\xb6\x0eu\rM\xad.m\x1d]=\x06\x95\x9e^}\x03\xc3>#c\x93\xae\xae.\x13S3\xa0@\xbf9\x93E\xcf\x04K+\xa0\x80\xb5\r#H\xc0\x96\xd9\xae{\x82\xbd\x83\xa3\x93\xb3\x8b+#\x83\x1bT\xc0\xdd\xc3\xd3\xcb\xdb\xc7\x17!\xe0\xe7\x1f0qR`P0L`rHh\xd8\x94\xa9\xe1\x11\x91\x0cn\xd3&\x80\x04:\xda{::\x80D\x14C\xf4\xf4\x191\xccv=Sf\xce\x9a\xdd\xd1>gV,C\xdc\xdcy\xf1,\t\xf3\x17$&%O\xe9MIMcH\xef^\x98\xc1\x9a\xb9hqVvN\xc7\x92\\6F\x86\xbc\xfc\x82B\xf6\xa2\xe2\x92\xd2\xb2\xf2\x8a\xca*\x0eF\x86\xea\x9a\xda:\xce\xfa\x86\xc6&\xae\xe6\x96\xdaVnF\x00\xbf\xc7S0N\xab\xe7\x10\x00\x00\x00%tEXtdate:create\x002016-08-30T01:21:10+02:00^(\x80!\x00\x00\x00%tEXtdate:modify\x002016-08-30T01:21:10+02:00/u8\x9d\x00\x00\x00\x19tEXtSoftware\x00www.inkscape.org\x9b\xee<\x1a\x00\x00\x00\x00IEND\xaeB`\x82'
+trayicon_away = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00 \x00\x00\x00 \x08\x06\x00\x00\x00szz\xf4\x00\x00\x00\x04gAMA\x00\x00\xb1\x8f\x0b\xfca\x05\x00\x00\x00 cHRM\x00\x00z&\x00\x00\x80\x84\x00\x00\xfa\x00\x00\x00\x80\xe8\x00\x00u0\x00\x00\xea`\x00\x00:\x98\x00\x00\x17p\x9c\xbaQ<\x00\x00\x00\x06bKGD\x00\x00\x00\x00\x00\x00\xf9C\xbb\x7f\x00\x00\x00\tpHYs\x00\x00\x0e\xc3\x00\x00\x0e\xc3\x01\xc7o\xa8d\x00\x00\x00\x07tIME\x07\xe4\x06\x01\x0e$\x12~2\xdc\x15\x00\x00\x07MIDATX\xc3\xbd\x97klT\xd7\x11\xc7\x7fs_\xde\x17\xb6\xd7\x80c\x83y\xd8\x06b\xea\x16B\xda\x08\xd7\xadhDb\x14\x02U\x03\x86\x14\x9c\x86FM[\xa9Q\x89\xd4\xa0J\xed\x87V\x15\xfd\xd4\xaf\xed\x97VU\xbf@\x1aLR7\xe1\x91\x80\x13\xf5\x05m\x02\xc14%\xc1\xc5\xbcL06\xac\xc1\xe0\xdd\xf5\xaew\xd7\xf7\xdes\xfa\xc1\xbb\xc6\xc6\xe6\xa5\xa0\x8etuuGsf\xfeg\xe6?\xe7\xcc\x95\xb6]G\xb8\'\x110mA+\xcd\xba\x8d\r\xb75\x8b\xd7\xd55\xabpx\x9d\x18F#\x99\xcc\x1c\x82\xc1KZ\xa9\xf7\x8dt\xfa\xcd\xd2\xae\xae\xb6In[_\xeb,w\x99Vt\xa7\xd8\x1aS\xb9\x84\x87\x96\xc9/\x92\xdd\xb4`i\x8dB\xb3\xbe\xe5&\x90\xab\xf5\xf5\x15N0\xb8SG\xa3\r\xf6\xb7\xb7D\xac\xc5\x8b1\xe7\xce\xc5\xbdx\x11\xff\xd4)\xbc\x1d;S28xd$\x93y\xbe\xbc\xb336\x06\xa0\xbd\xf5\xf5\x13\xa5\xc6\x7f\xa3\x82\xbe-\x00\xa5M7\xc7\x8c\xde\xa4\xaey/\xad\xe7\xbe\x16\x91\xee\xee\x8c\xae\xc2\x94\x1c\xeb7-\x1f\rn\xdb\x9d\xb4l\x8e\xa8\x17^pN\x9e\xed\xe2\xfa\xc05\xd2\xe94\xc1`\x88\xb2\xe9Q\xea\x17.\xc6\xde\xb1sDv\xb5\xa6F\\\xb7\xbe\x00\xc2\x9a&\xe7\x97<l\xfe\x01\x03\xff\x0e\x19\x10<"5q\xb5hE\x9fjzv@=\xf6\xb3\xa3\xb2t\xefr}B\x03\xd8\xc1\xa2\x9dj\xc3\x86\xc8\xe0\xc6u\xce\xbf\xff\xf1W\x96\x16\x97\xf2HI\x19N\xc9\x0c\x06\xeb\x16\x12\xeb\x8f\xf1\x97C\x7fc\xc93k\x9d\n\xe5G\xec7\xdav\x02M\x00\x86\x00\xa3\xbb\xbf\xfd#(l\x12\x94\x1bGYl\xfe\xee\x0bU\xe6\xc1\xdf7\xea\xf77\xd7\x19\xbf%^W\xd7\xacK\xa3\r\x99o=\xe7ttt\xf0\xde\xbe\xfd\xc8\xf9\x0b8\xe7/\xa0\xbb\xba\x08\x04\x03\xcc\x993\x87G\x96-\xe3\xe3\xff|\xc4p\xcb&G\x95E\x1b\xe2uu\xcd\x00\xc6\xbd1\x90<L\x93\x80\x0cPk\xb6\xce\xac4\x0e\xfd\xaaK}\xff+*\x1cZ\xe7~sc\xe4\xd3\x9e\x1ef\xcf\xaa\xe2\xf1\xd5O\xf3\xca\x9e=\x1coo\xc7;|\x18\xd7uq]\x97P(HE\xe5,N\x9d=\x83\xf7\xec\x86\x88\x17\x0e\xad+\x00\x98\xa2\xf8\x1a\xc1\x9f\xf0\x14\xcc4\x06\x0eq\xe6\x99{\xaaJ\xe5\xd4O\xb5a~\xd5\xaf\xa9&\x9eL0w\xde<\x9ex\xf2\t\x9a[6\xf1\xca\x85s\xfc\xc0\xcbr\xe0\xe0\x01\xb4\xd6\x98\x86IEy9\x89d\x02\xaf\xa6\x1aC\x8cF\x00CO\x11<\xabg\xd2\xa7V\xd1\xab\x9e\xa2W=E\xbfj\xc4#4\x01DD>\xa5\xdc\xf8`\xa5\x91\xcdV\xf9\x95\x15d\x86\x87)\x9eVB8\x14f\xcd\x9a5,\xa8{\x98\xc6\x95\x8f\xb3o\xef>fUTQ5k\x1e\x91\xe2br\xd9\x11\xbc\x87\x1eB\xb2\xd99\x00\xd6X\x93\x8f%Z\x91\xd2\xf3\xe9\xf2\xbe\xd7\xea\x13|\x07\xc0$\xf3\xf9\xf9\xe6\x9f_\xaa6\xdb"\x92\'\xab\xa0)3>\t\xba\xa5\xd1\xa4\xbe\xd4[\x1c\x89D\xc8e\xb3\x04JFA\xac^\xbd\x9a\xf6\xf6v\x9aV5\xb1}\xfb/Q\xcag$\x97#\x18\x0c\xc0\xe5>t p)\x0f@\xf4\x04\x04\xf9}\x1axG\x85\xa1\x9d.\x11z\xf4\x8a\xa2\x90\xbaR_i\x1cZ\x13\x92+h\x04\x8dP\xc4uR\xe5\x8bR\xf6\xf9s\xc5\xa5\x8f~\x89X\x7f\x8c\xb2\xe9\xd3\x11\xa5\xd8\xd0\xbc\x91\xf5\xeb\x9aQJ\xa1\x94\xc2\xf7=N\x9f9CiI\t\xc6\xa9.\x94R\xef\xdf\x8d\x84\xa6C\x02A\xb35$9\x9f\xa2\x98O`\x02e\x04E\xaav\xce@\xe0Oo\xa5\xca\xca\x8a\xe9\xe9\xb9H<>\x88m\xd9\xd8\x96\x8dc\x17\xe18E8\x8eC:\x9d\xa1\xb7\xef\x12\xb3++\x08\xb4\xbd\x95\xb2\x86\x87\xdf\xbc\xa7.h\xde\xdc\x00\x81B^&3f\xe0\x8bK\xaf\x9a\xc9\xd4\xb1\xe0\x1fw\x8f\xcc\xaf\x99\xcb\x91#\x1f\xd0\xd9\xf9\t\xf1D\x02\xcf\xf3\x18J&9s\xe64\xc7:\x8eP]SC\xb8\xf5\xf5\x11+\x9e8R8\x96-\xb9\x1b\x82{\x90\xcb\x15K^\x9c\xfd\xd6\xbe\x0e\xadu\xa4v\xd3Fg0\x91\xe4\xf2\xb1\x0fI\x0f\xa7\t\x86\x82D\xc2a\x16-\xa8e\xda\xee7F"{\xdeN\xb9\x9e\xf7|a\xad1\xf5\xbe\xeeO>\xd8\xfc\x9b~\xd7\xf3\xea\xa7\xed?ph\xc6\x0f\x7f4<\xeb\xc3\x0f\xa9\xce\xa5Y\\UAm.C\xd5\xf1\x0ef\xbe\xbc-\x1by\xfb\xdd\x7f\xba\x9eW?\xfe.\xb0\x1e@|\x00\xf2N\x9bN\xbc\xb4\xed\x8d\xb2\x8e\x83\x1b"{\xe3X\xd72x3Cx\xb3l\xae4<\x968Y\xbfu\xcbs-\xb5\xb1\xf1\xeb\xac\x07\x11|\xbc\xf4\xaeXy6j~D\xd0\xb8\x80\xc6\xc4B\xa3\xb0I{k3\xbe\xefL\xaa\xf8}\x1c\xc5w\x96\xb6]G\xf9\xa8u\x1b@\xbeMM4\xc6\xd8\xbb\xd0\xe9\xb7\xce\x1f\x0f"\x03:MPUq\x99\x94\x9e\x9f\xbf\xd9\xa6\xa4\xb6V8~\x80\xeb\xbc\xbbk7 d\x98\xf1\xd98 \xa2\xd1Z\xaa\xe7\x1b\x87~\xecJ\xb1{C\xd7k\x8b\xe1FAM\xb4\x1b\rQZ"g^\xf6%\x10\xf7u@\x00\x14\xf6\x80u\x1b\xb4\xf7\x12\x1e\x8b\x14\x0b\xcdW\x17x\x84\xb6\xdf\xd4z\x84\x8dK\xe8[\x8e\xf7\xb9\xe6\xfeh\x85\xf1\xaf\x9f\x8c\xa5\x03\x83\x1e\xf5\xb4?\xe9.\xb8\x1f1\xf0\x88\x1a\'\'%q|\xcd\x0b\x9a\x129\x9b?\xf5\x0b\x1a\x8b\xabj\xb9\xf9\x999\xa0\xef\x91\xc7\xb7\xda\x15@\xdef\x1e\xf8\xff\x89Q\xc0#c\xe3\x97\xbe\xcf\x82\xe8\xbb|\xdfYo\r\xe9\xea\x91\xb3\xfe\x16\xfb&s5\x19]\x89\x8f\x93\xd1\x94\xde\xe2\xa2\x00O\x00\x8dG\x90>\xffI\xd7\xa58)(\r\x82\xe0\x87+\x8d\xbf\x07\x83\x12\xa3\xc0\x03\x8dAL\xadPi]\x95\x10|?\xaf\x93!=?iY\x0c;a\xe9\xe3&\x00\x85\xc2AP\x96\xe0\x8dj\x86\xc1\xd5\xd3t\x9f\xbf\nG\x06\xc7\x1c{\x04\x89\xa9\x15\xc7R\xbaz\x9b\xe0\xe7\x14\xa6\x0e\xd3\xbb5*\'\xbfS\x98\x1b\n\x00\x06\xd4\xa3\xb1\x8b\xea\x99\x17Mr1\xd0\x02\x88\x81;d\x85\xa4\xcf\xae2\xda\xb99\xe9\xf8\\U\r\xc4\xf8\x9am3\x04@\xaf\xdf\x84\xc6\xf4G(Ek\x13\x8d hF\xe7\x03\x89g\xf4\xf4\xe3\xc5\xd2\xed\x16\xd3\xcd\x88\x94\xc4\xa6J\xb7\xe0\xe7\\\x1d>Q$\xd7\xafL(Aa\xda\x9d\x98jcl\x19\x80%Y@\xcb\xe8\x80\xea!\xf9\x12\x8cfH\x8bE\xda\xd4\x1aw\x86q\x9c\xcb\xac\xbc\x1d\x85D\xf0M\xc8\xcf\x18\x13\x01H\xdei!\xa8\x8c\xda\xe7\xa5R\x0ec\xe0\x1bEr\x83"n\xe4\xedGK`h\xd7T\xda\x96\t\xe7\xce\x04\xae\x8c\xf9\x9c\xb2_\xad\x84^\xf0\xf1i\xef\xbb\xd1B\xda\x04M\x86\x99\x9e\x8f\xddS \xe1q\xef\xe7h\x8c\xee\x9c\x8e\xf6*,] \xa1O\xc0\xd0\x98\xe7\x94\xd8\x1e\x02\xfd\xba\x11\x8f\xd0\x85\x1e\xb5\xb6\xa7_7\xca8\x12\x1aI]\xd3iK*3\t@\\\x7f\xae\xe9\x9aZ^4\xbe\xf9Lr\xaa\x88\xf8\x80/A\x00\xba\xfd\x8d\x18x\xbf\xbe\xe0oxU\x8d\xbb\xbf\x04\x1fG\x92\t\x8b\x11\xd7\xc6a\x88\x1a@\xed\xe8\xf1\xbf\xfe\x8e\x8f#\xe3\xfe7\xc5\x96\xa1t\xb1y\xf6FV\xcf\x98\x08\xc0f\xe8\xaa-C\x93\x0b\xa6\x05\x9dw`I\n\xd1\x92\x0eJ\x7fz\xca4J\x9aol\xfe2m\xbb\x8e\x02:\x1b\x90k\xbdS\xd9eu9\xb7\x12\xf4\x7f\xfb\xd4F\x945\xc8\xaaT\x00\x00\x00%tEXtdate:create\x002020-06-01T14:36:18+00:00\t\'//\x00\x00\x00%tEXtdate:modify\x002020-06-01T14:36:18+00:00xz\x97\x93\x00\x00\x00\x19tEXtSoftware\x00www.inkscape.org\x9b\xee<\x1a\x00\x00\x00\x00IEND\xaeB`\x82'
+trayicon_connect = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00 \x00\x00\x00 \x08\x06\x00\x00\x00szz\xf4\x00\x00\x00\x04gAMA\x00\x00\xb1\x8f\x0b\xfca\x05\x00\x00\x00 cHRM\x00\x00z&\x00\x00\x80\x84\x00\x00\xfa\x00\x00\x00\x80\xe8\x00\x00u0\x00\x00\xea`\x00\x00:\x98\x00\x00\x17p\x9c\xbaQ<\x00\x00\x00\x06bKGD\x00\x00\x00\x00\x00\x00\xf9C\xbb\x7f\x00\x00\x00\tpHYs\x00\x00\x0e\xc3\x00\x00\x0e\xc3\x01\xc7o\xa8d\x00\x00\x00\x07tIME\x07\xe4\x06\x01\x0e$\x12~2\xdc\x15\x00\x00\x07vIDATX\xc3\xbd\x97kpU\xd5\x15\xc7\x7fk\xefs\xce\xbd7O\x02\t4\x10 \x84\x87\x9d\x02*0\xc8\xa3J\x15Dg\x9c:-*\xa2X\xec\x88\xd4\xceX\xec\xb4\xd3\xa1C\x9dJ\x19g:}0\xb5-\xfd\xd2qJ\xa1\xea\x18\x9e\xedH}Q\xfb\xe0!\x03\xd1""\x04\xd1@\x84\x90\x84$$!\xb977\xf7u\xce\xde\xfdp\x93\x90@@\x18\x99\xae\x99\xf3\xe1\xecYg\xad\xff\xfe\xaf\xff^{\x1d\xd9Qy\x90k2\x01\xed\n\xd6X\x16-\x9e}E\xb7iOP\x82\xd2\xf7\x8bp\xbbV\xfav\x04c\x82\xe0\xdd\xc0\xf0nF\x82\x9d\xd5\x1bh\x1f\x10v\xf3\xab\xd5\xc33\xe4\x87\xae\x96\xdb\xa2M\x86\xdc\xd84Y\x1b\xade)\x8e\xb5\x18,\x0f,\x1d\x08d\xfa\n\xfd\xc0\xb0\xc2\xd2\x8d\x0f\xdd\xb9<<\xb9b\xa67\xa9l\x1a\xae\xf6\xa8i<\xc2\xb1\xda\xf7R\xaf\xbc\xbd>\xd1\xd1\xd9\xba\xec\x83\x8d\xc1\xeb}\x00vm\xdezd\x88:^$\xd8+\x020VgR\x14\xd7Gm\xc5;q;\xe6\xd5<\xa9\xadM\xd82\xb4\xa4x\xe0\x91Y\xd9\xe4\xcb\xf5\xc69S\xef]\xb2\xe4\x9e\xa7"G\x1a\xfeMm\xeba\xea;O\xa0\x95b\xec\xb0)L(\x99\xc9\xad\xa3\xe6\xb3q\xe7\x0b\x89\x03\xd5\xefl:\xbc!x\x1a\xc0\xc9\x97S7\xdf\xa47\xa0\x08\xae\xc2\x80\xe0\x93W\xd1a&\xcdk0\x0b\x1fn53\x9f\xab\x92[v\xce\xb2G,\xc0\xf4\'\xf4\x92\xd9S\x17.\xb9c\xe6\x82\xc8\x8b{V\xd1\x91n\xc0\tI\xb6d"\x9c\xbcP\xc5\'-\x079p\xea5\x1e\x9a\xff\xe3H(\xe2.\x1f1~w\xfc\xedgc\xab\xf4\x8a\xc5w\xaf\x1d\xae\xdeC0W)\xbfE\x93$O\xea)\x92\x8fG\x18\xf1\x16\x14\x1a\xaf\xe1&\xf5\xe7\xa3\xfb#\xff\x1d^T4\xe2\x9f\x8b\xeeZ\x96\xbb\xfd\xfd\xdf\x12\xcb4\xe3\x84\x14nH\xe1\x86\x15NHPZ\x00\xa1+\xd9\x81M\x0bKo_\xed\x1ch\xdc2g\xf4,\xb5]]\x9b\x02/\xc2\x08K+\xe3\xf5\xe6\x92R\xb5\xf7W\'\xccS_\xd5\xaeZ6w\xea\xc2p\xd5\xc9\xb7\x88&[\x11\x95M\xa8\x1dp<\xc1\x8bh\x1cOP\x0e\xcc\x19\xb3\x88\'\xe7\xfe\x86QE\x13\xb9y\xe4\x9d(O~\xa1`\xb0\xe2[\x84`\xc0\xd3\xebfQxt0V\xbfV6D>\xfe\x89\xa3\xbd\xf9C\x0b\x86\xb9\xb5\xcd\xd5\x17#Y\x98\x94w\'E\xeeh\xac\xb5\x08\xc2\xbd\xe5+y|\xc6\xaf\x89\x84rp]\x97\xfb\xa6>%\xe1\x88w\x9bc\x07I\x9e\xb4%\xb4\xdb[\xb0d\tr\x892T}\x84Cw\x0f\x13\x8a<9\xcdpu`~\xd8\xcbM\x06\xd6\xa7\xb3\xab\r/G\x81\x11\xee(Y\xc1\xec\xd2\x87I\xa9\x0e\xb6\xd5\xff\x90\x05cV2\xabl\x11\xae\xeb\xe28\x0e\xae\xebR\x9c7\x86\xd8i\xcfw\xfa\x0ey\x1f\xd1\x86.[\xce\t\xff;\x9b\x03"o\x02h\x12S\xca\xf5_\x9f\x1e\xa7w\xe4I\x8fX\x05\xcbPu4\xa2E\xa9\xae\xeeNl\x00\xc6\xb7\xdcV\xb2\x84\x19%\x8b\xd0Z34<\x8a\xefM\xdeNn$\x1f\xadu_\xf2]\x1f\xfd\x85O\xcfV\x13O\xc6\xc3\n\xc4^\xce\x01(\xfc*\x87\xd8\xcb\x16\xfbr\x9d\x9d\xb7\xa6\xc5\xcc\xd9\x93\xb4%H_)\x84\x10m\xe4{\xf9]\xe9L\x06WE\x082\x96\xfdu[9\xd5Y\x85\xd6\x1a\xad5!7\xd2\x97\\)a\xd3\xde\x9f\xf1\xc7\xb7\x9e%\x1a\xbb@*\x1d?t5\x11j\x8fN\x04\xcb39\x92\n\x085\x05\x84\x07HF0\x94\x17Tt\xb6\xc7\x9b\x83|o\x18~\xda\x92N\xa6\xa9\xfcp-\xc7Z\xfe\x83R\n\xa5\x14Zk\x02\x9bf\xdd\x1bO\xb2e\xdf\xef)\x08\x95\xd0t\xfel\xc6\x18\xbb\xfbsO\xc1\x83\x8f\xce\x86p//\x97+f\xfa\xc89\xcd5\xa7\xab\xd3\xe5\xa3&b}M&eI%|^:\xb8\x86\xf7\xcf\xbc\x81R\x8ah\xb2\x8d\xe7\xb6=\xc8\xde\xa3\x7fG|\x8fQ\x85\x93\xa89[\xedc\xcdvG>\x0f\xc1\xe7\xd8\xd8\xc2q\xf1\xeeD\xe2w\x8d-u?\x98T>9r\xa2\xf6\x18\xc67\x04\x19\x9f?\xed\xfe)m\xf1s\xec;\xf17Z:\xea\xc1h*J\xa7p\xa6\xe9Twg\xbc\xed\xf9\x0f7Q\xa3\x06\xdf\xd7\xf5\xd9}3^\xfa\xe5\x99\xb3\xa7j\xbb\xe3\xf1\xeei\x93gPTPL\x90\x14\xd2\xdd\x01[\xf6\xac\xa7\xe9|#\xb9z\x18\x93\xc7\xce$\x1a\xef\xe8>}\xee\x93\x13\x87G\x9bu\x00\xce\r\xc8\xcf\xc8\xdc{}}!=\xf3$5\xebZZ\x86~\xb7|B\x99S>z<\x99 \x05\x164!\xba\x13]||\xf2x2\x9e\xe8|1\x1a\xf1W\xb36\xdbz\x9d/\x9a\xbc\xd7\x0el#\x01\xc1\xca\x9f\xbf\xf0\xed\x82\xb6\xb6\xcae\xa7\xa2m\xb4\'\x1d\x10\x18\xe2\xf9\x8c\xc8\xfdrWq\xce\xad\x8f\xed\xfb\xc3\xbfv\xf6\xff\xee\x86\x01\xd8QYE\x85l\xa5\x91\xbb\xea\xa7\x8c>\xccPu\xac\xaf\x91\x19\x1c\x8e\xfb\x8f\xb4\x9e\x0c\x1e\xfb\xe0[\x95\rYa\xdf@\x006N\xc4\x94\xd1H\x97-\xcfv(\x06\x95\xb65xA\x986\xfeQ\xb9\x05\x10\x12\x14\x7f1\r\x88X\xac\x95q\xe5j\xef\xaa\x8c\x14d\xda\xedd\xeb\xd0=\xf7\xd2\x9b\xb5\xa7y\r)\x94O\xbf\x1fH\xb8#\xb0a\xc92\xe3\xb6:W@{-\xe9q\xe8b\xa2~e\x82O\xce\xf3\x17W}r\xd5Y\xec%\xed}\x8c~\xbd\xe8Kj\xff\xea>:P\xd4\x99\xfb\x82\xcb\xee\x82\xeb1\x85O\x91:v\x19\x89\xd9\xda\xcb\x80\x95B\xa9\x19\xd0\xf5-\x0e-f\x96\xfe\xc2\x1a\xe8\x15\xda\xf5\xfa\xf5\x82\xbc\xc2<\xf0\xff3\xd5\x8bGzz\xbd\x0c\xa8\xde\xb5\xed\xed\xea\xefW_wbv\\\xba&x\xdc\xbd\xa8\\K\xc2\x96\x12\xe0%,C.\t\xd1\x0bO\x00\x8bO\x84\x86\xe0\xeeL\x86\x82\xa8`,\x08B\x90[\xaavG"\xd2D\xaf\x0e,\x8a&3\xcf\xc4mY\xa7\x10\x04=k\x12\xb3\xe5Q\xc7\xa1\xdb\xcb\x95\x86~C\xa9\xc1\xe0!\x18G\xf0\xb3+\xdd\x90\xb1\xf9\xb6!\xb8\x07O.\xf4\x05\xf6\x89\xd0d\xe6\xbd\xdfe\xc7\xfdH\x08R\x06ms\xa9\x7f\xa6H\x8e-\xcf\x91s}\\Z\x14\xadfz\xd3\x19\xf3\xcd\'5\xa9&\xb0\x02\x88"\x13sr\xa4\xc1-S\xbb\xb88\xe9\x04\xb4\x98\xd94\xf15\xd7%\x06@}\xb0\x10\x8b\x0e\xd2\x0c\xc1ZMv\xca\xb3d\xe7\x03\xe9H\xd8a\x87\n\xa46S@-i)l\x1a\x8cn!Hel\xee\x91\x90\xb4\x9d\x1bP\x82\xdeiw \xd5\xaa\xef3\x00G\x92\x80\x95\xec\x80\xea#=%\xc82d\xc5!\xae\xad%S\xac\x0e\xd1\xc8\xfc+IH\x84@\x03\x97\xb6b\xe9\xd9\x91\xf4K*Y\xff\x1e+\x95}(\x02\x15\x92vB\xb4\xf7\xf8gK\xa0lF\x1b\xeb\xca\x80\xbe3@+}1\x07=\xafN\xa7\x9d\xf0\xd1\'\xfe\x8a\xa2^\xda\x04K\x82\x12?\xc0\xad\xeb\x15\xe1!\x7f\r\x16U\x9b\xb2E\xf5\x06\xc7\xf6\x8a0 \xac,\xfa\xa4\x11\xd7G\xa0\xd9\xce\xc5\'\xe7\xb3:\xf3\xf5\xbaf;W\xfa\x89PEmE\xb5+]\x89\xcb\x00t\xd8\xaf,<of\x85\xfa\x1f>M\xca\x84\xe8h\r$\x02@m\xb0\x18\x85\xbf\xfe\xb3\xe0\xa1WL\xbf\xfbK\x08\xf0$\xda\xe9\x90\xce\xb8x\xc4\xa8\x00\xccKu\xc1\xfdo\x06x\xd2\xef\x7fS\\\x89\xc5\x0btM{\xd2\x16\x0f\x04\xe0\x12kq%vy\xc1\xac`{\x028\xd2\x85X\x89G\xa49>(\x8d\x12\xe7\x1b\x8f\xceaGe\x15`\x93a9_?\x98_\xd2\x0e\xe7R\x81\xfe\x0fO\xf9=\xb3\x89GCa\x00\x00\x00%tEXtdate:create\x002020-06-01T14:36:18+00:00\t\'//\x00\x00\x00%tEXtdate:modify\x002020-06-01T14:36:18+00:00xz\x97\x93\x00\x00\x00\x19tEXtSoftware\x00www.inkscape.org\x9b\xee<\x1a\x00\x00\x00\x00IEND\xaeB`\x82'
+trayicon_disconnect = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00 \x00\x00\x00 \x08\x06\x00\x00\x00szz\xf4\x00\x00\x00\x04gAMA\x00\x00\xb1\x8f\x0b\xfca\x05\x00\x00\x00 cHRM\x00\x00z&\x00\x00\x80\x84\x00\x00\xfa\x00\x00\x00\x80\xe8\x00\x00u0\x00\x00\xea`\x00\x00:\x98\x00\x00\x17p\x9c\xbaQ<\x00\x00\x00\x06bKGD\x00\x00\x00\x00\x00\x00\xf9C\xbb\x7f\x00\x00\x00\tpHYs\x00\x00\x0e\xc3\x00\x00\x0e\xc3\x01\xc7o\xa8d\x00\x00\x00\x07tIME\x07\xe4\x06\x01\x0e$\x12~2\xdc\x15\x00\x00\x06XIDATX\xc3\xad\x97]l\\W\x11\xc7\x7fs\xee\xddO;\xeb88\x01KV\xc8:M\x81\xb6!\xa5\x05%\x84\xe6\xc3mSU|\x08h\xa9D\x8a\x94\xa26\xe1\x01Q\x1e(H\xe9\x03\x11\xca\x0bA\x02\x1ex@\xa06E-m\xed"\x05\x89\x8aV\x94\x086\x1f$\x10E)J\xdc@\x82\xdb5u\xe3\xc6I\x93`\xefzc\xaf\xef=gx\xd8{\xd7\xde\xb5\x938MG\xbaZ\xed\xd9sg\xfe\xe7?\xff33+{z\xff\xc1\xbcL\xc0K\x08\xea\x94\xaf>\xb4f\xce-\xbbv\xed\\i\xfc\xe4!\xe7\xec\x82\xe6\xdf<\xcf\x94l\x10\xdc\xb5}\xfb\x8e\xfe\x06\xb7}/\x9e\\\x12\xb0 u\xb5\xd8\x8a\xe7\x02Z\xca\x9f\x92\x1f\x95\x8a<\x8c\xaf\x8aCy\xe0\xe15\r\xc1\xfdd\xea`\xcf\x86\x9e\\>\xdf-\xcd>\x06\x07\x8bZ\xd8W(\x85Au\xddL\x10\xf2Z\xdf\xef\x8e/4\xffj\x17\xf4\x8a\x00\x9czA\x95\x8e3%\xed\xde[\xd1\xa5/\xb6J\xb18\xa1]xR\xe5\x81\xaf\xafn\x08\xderz@^\xdf\xf2\x08.\x0c\xeb\xef\x1b\xdf\xe7\xce\xde\x17\x18\xef\xceka\x7f\xa1\x14NM\x83\x90\xc3/\xfdXo\xf1\x7f\x89\xc1^\x85\x01!\xa4\x95Qw3\xc3nS\xff\x05\xf7\x99\x1f\x1e\x965/\xaf\xd6\xe3\xfa\xd6\xdb/\xaf\xf4\x13\xa9\x83=\x1b{r-\x03oI\xff\xd6m|\xba\xa3\x83V\xdf\'\x08\x02\xc2 `lb\x82\x7fOM\xf1\xc9\xddOSY\xb1\xbc\x01\x84\x11\xa0v\xfa+?\x82#\xc1\x18K\xcc\x11>\xe1\xfdze\x97\xf7\xa7\xa7\xd6\xea\xe1\xcd\x1f7\xbf\xc2K$\xfe\xd6\xb3\xb1\'\xd7Z\x1c\x94\xfe\xad\xdb\xb8\xbd\xad\x8dV\xdf\xc79\x87\xaa\xa2\xaad=\x8f\x15"\x9cxl+-\xa7\x07\xa4gCO\xce\xf8\xc9C\x00f~\n$\x82\xe9\x91\x96\x0b,\xf7\xfa\x16w\x9a\x03?9\xe5\xbe\xf59k].\x9f\xef\x96c\x9b\xbf\xc1m\xd9,Y\xcf\xab\x9f\xdcY[\x03\xe2\x1c\x19c\xc8\x87!\xafoy\x84|\xbe[b\xa1\xfa\xd11\x9bDS;u\xe3\x8a\x89@\x18\x92\x8c\xf2Q\xef\x0f]\xe3\xba\xf4\xc9\xbaN\xc2\x90\xb4\x08a\x10 \x80\xaa\xe2\x9c\xc3Z\x8b\xb5\x16g-\xa9h\xdfL\xf3gKO\x99\xd4\xc5\\\xd2UQPHPb\x919\x81\xcf\xe5:\x88V\xf9/K\xcc\xdf\xef\x86\xce\xfa\x9ba\x10`D@\xa4N\xbf\x8b\x00\xc4 \x9a\xcd\x8f\xe9\x9d&\xda1\xae\xcb8\x15n\xeb\xb3d^\x05\xf0\x98\xb8m\x99\xf7\xfbo\xe7\xbd=\xad\x12\x89UP\x16\x99\xfeL3\x00\x89\x00\x10\x03p\xae\x06"\x0c\xb1s\x03\x90Y)P\xc0\x10\x1e\x11\xca\xbf\rheH\xd7\xa7\xb2\xee\xec\xad\x9d\xe6\xc0\x17\xb2r\x16EP\x84\x14\x17\x1b\xaeZ\xb9Z%\xe3y\xb5\x14Di\xd0\x18\x80\xb5LF\xfbf\xda\xd5D\xe8%\x19CP\x1e\xcfJ\xd5\x92\x1a\xb1\xa4#\xd7\xd3l\xf9\x9eo\x07\x07\x8bz\xc7s\xcf2\xe8\xfb\x8cW\xab5\x11NM\xd5\x9e \xa8\x07\x1f\xcef\xb9\xb3\xf7\x05\x06\x07\x8b\xeay\xa64\xaf[\xf0\xe0\xe65\x90\x8ey\x99\xad\x98\x9b\x17\xb7\x1d-\xec/\x94+\x1f[\xa1\xab\x9e\xd9\xcdP:M%\xa2;\xbe\x8aq\xf0U\xcf\xec\xae\x15\xa3}\x85\x92\r\x82\xbb\x00\x8c\\\x0b\xc15\xac5\x95\x1cW\xd7roa\x7f\xa1TY\xb1\\o\x7f\xf67\x0c\xf9>\xa7\xa0\xfe\x0c\xf9>\xab\x9e~j\xba\x08\xcd(\xc7~|\xae\x1b\x01\xb2\xb0\xf3\xd1\x93\x17\xdf\xf9\xe9\xba\xc2\xbe\xc2\xc1\x9e\x8d=\xb9\xfb\xcf\x8f\xcc\xdd\x0b\x9a\xca0L\xd7\x81\x1b\xb6\xed\xdbw\xf4\xef\xda\xb5s]a\xff_\x8f\xee\xfd\xcb\xdeY\xcd\xcd\xf7|M&>\xfc\xe0\xf7\xbf\xf7\xcd\x86n\xe8\xcf?\xc4\xfc@\xbc\xd2\xf7\xca\xcfWz?{r\x919\x81\xe2\x01\x8a#\xc1\xc9\xf0;Co\xda{N7\xbfs\x1d\xa5\xf8\xea\xb6\xa7\xf7\x08\xff\xec{\x02 \xba\xa6\x1e\x8a\xa9\x7f\xc6In\x9e?>\x08\x06\xb4B\xc6u\xf1.\xe3\xba,\xeals*J\x1dI\x9b\xe6"\x7f\xee}\t\x10&\xe8\xb81\r\x88(\xaa\x92_f\x0e\xfc \x90\\pIoU\x9f\xcbk\x9b\xfbH4k,l\x93\xff|\xd7Jz\xd4jZ\x00\x1c\x89\x0b\xfe\xfb\xd7\xbf\xe03\xce\n\xef\xf9\x9bB\xb2;\xa7WCZ\xcc;hSy_\xea\xfd\xb1\xfd#\xe6\xd0\xf6:\x1d\x18\x86\xdc\xe7\xed\xac^p=f\x08i7o\xcc"qf\xce\xe3\x956\x19\x88\xaa~\xbc\xe2s\xde\xad\xf6nX\x03:O\x1d7\xef\x8bA\x1a>\xa0:\xf0~\xcd\xc4x\xa4>~\xe9u&D\xaf\xf1\xfd\xea\xeb~Y\xf3S\x03vKbZ\xb9\xca\x84vbIN(\x0b\x9b\\\xc4\xf0j\r7$\xc3\xb0\xbd7\x08\xc8\x95\x04\xa7 \x08\xb6\xa5\xd3\xec\xcbdd\x84X\x07\x8aa\xc4\xadw\x15\xed\x1a\x13jC\x81b\xa4\xac\xcbJ\xbe\xcf\xe5d\x8b\x0c\xcf\x18\xc1\x1c\x8e$\x82\xf3\x85\xda\xf8\xe4.C\xa0\x0bt\xd8\xdeGR\xfeWw\x1c\x92a\xc4\xad?:\xae\xf9\'\x04[ux\xda\xc2\x99\xc7\xdb\xe5\x8dG\xe3\xb9!\x06p\xc1\xdd1\xf2\xb6\xfb\xcac\x1e\xd5\x11P\x01\xc4\x10\x94\xfd\xac\x0c\'\xba\xcckLO:\x96\xf3n\r#lH$(\x03p\xc6nB\xf1\xec\x14\x0bQ\xf5P\x04A\xa9\xcd\x072:\xa1\x1f:\x96\x93b\x90\xa3\xc8\x94\xb4\x8d\xccE\xb7`\xab\x81\xb6\x1cO\xc9\xc5\xb3\r)\x88\xa7\xddF\xaaM\xfd5\x00_&\x01\x15\xc1"\x84H\x94\x82\x1aC*>\x15O\x95\xa0\xc3\x1c\xe3]\xee\xbe\x92\x84D\xb0\x1eD3F#\x00\x89\x9c\xc6A\xa5\xb6?\xb2N9\x88\xc1\x9a\x94\\"\xc5\xa5h\x7f-\x05F\x03\xcfiB\x1a\xeaN\x83V\xea>\xe7\xbc\xaf\xfe\x98\xdet\xe2t\xb8\xb5=\xa6MP&X\x1cZ\x12C\xb1\x08\x8f\x85;PL\xb1\xaa\xedg\x1c\xbe\xc6"\xb4\xa4\x8d\xe2\xbd\xe9$\x11"pN\xd7\x12\x92\x1d\x1cr_\x1c:\xa7ke\x86\x08MI\xbbO&d|b\x16\x80Q\xbde\xd3{nuj\xe6\xe5\xf3\xa8\xba\x14\xa3\x17\xacd\x00(\xda\x870\x84\xbf\x18\xb4_{\xde\xcd\xe8_\x82%)\xa51\x9f\xa9 A\x922\xdd\x80{n\xc8~\xe9UKRf\xfc\xdf\x94\x84\x94+9o\xe0\xd2\xa4v4\x02HP>\x9f\x90\xf2\xec\x84\xa9\xa0\x91\x03_\xc6\x11\x95JF\xceU\xe6\xa4Q*|y\xf3g\xd9\xd3{\x04\xd0\xc9\xb4\xbcwf\xae}\x93\xba\x84f\x81\xfe\x1f\x01\xa2=\xcf\x88\xf5\x7f\x0e\x00\x00\x00%tEXtdate:create\x002020-06-01T14:36:18+00:00\t\'//\x00\x00\x00%tEXtdate:modify\x002020-06-01T14:36:18+00:00xz\x97\x93\x00\x00\x00\x19tEXtSoftware\x00www.inkscape.org\x9b\xee<\x1a\x00\x00\x00\x00IEND\xaeB`\x82'
+trayicon_msg = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00 \x00\x00\x00 \x08\x06\x00\x00\x00szz\xf4\x00\x00\x00\x04gAMA\x00\x00\xb1\x8f\x0b\xfca\x05\x00\x00\x00 cHRM\x00\x00z&\x00\x00\x80\x84\x00\x00\xfa\x00\x00\x00\x80\xe8\x00\x00u0\x00\x00\xea`\x00\x00:\x98\x00\x00\x17p\x9c\xbaQ<\x00\x00\x00\x06bKGD\x00\x00\x00\x00\x00\x00\xf9C\xbb\x7f\x00\x00\x00\tpHYs\x00\x00\x0e\xc3\x00\x00\x0e\xc3\x01\xc7o\xa8d\x00\x00\x00\x07tIME\x07\xe4\x06\x01\x0e$\x12~2\xdc\x15\x00\x00\x06\xa3IDATX\xc3\xb5\x97kp\x94g\x15\xc7\x7f\xe7y\xdf\xbdd\x13\xb2\t!\xa1\x94\x85\x86\x08\x05q*R\xc7rQ\xd2\xb1\x03|`@\x14A\x05G>\x94\xe2\x8c\x0e8\xb6\x95\xa9\x8ec\xed\xf8\xc1i\xd5O|\xb0\xe0e\xd4R\nu\xca\xb4\x8ah/\xe3h\x91R\xa2C1\x94"$\x18 $\x90{6{\xc9\xee\xe6\xbd\x1c?d7dC\x12.\xc53\xb3\xb3\xf3\x9e9\xcfy\xff\xcf\xff\xfc\xcf\xf3\x9cW\x0e\x1d8\xc1-\x99\x80\x15\x10\xd4W\xbe\xb0i\xe9--\x198\xc6R\x80\xe8g\x98\xf0%\xb6\xa3Sj\x1c\xa6\x84&K\xa4X\xbe\xa3\xa5\xc9\xc5C\xcf$Z\xd8\xc2k/5\xe0\xa3l\xd829\x101\xecBQ`\xe3\x841o\x1c\xfc}c\x859[)\xe8\x84\x89|\xb5\x9c\x1c\xd3\xda\x12Z\xf7VZg\xbfT&--\x19\x8daI\x8e\r_Y2\xee\x9a\xe4Q\xaa%\x12\xbb\xa2^\x02\x1dL\xc4\xca?K\xcf\xb8\x0cL\x91\xff~|\xbe\xf5k\x0c\xde$\x0c\x08.euq\xff\xfe\xfav\x7f\xd5\x97z\xfcO\xfd\xa0A\x16\xfdq\x896N\x88\xda\x83\xed\x91\xda\x1f\xdan\xe6\x14C\x97\x7f\xfe\x18\xf0\xecxqF\x80\xe1\xddO\xfc\x13|\x02\x0cPc\x1a\xf8\xa8\xb5\xf7\x81\x98\xf5\xfa/\x97\xeb\xf1\xcd\x0b\xcc\x9e\xf1\x01?\x83I\xe7\xcc\xb7\xed\xe8\x1a+P\xb9\xc9Jg\xcc\x13\xaa\xc8\xb8\x0cp\xcb&(\x16a\xe9\xe1#\xd6\xc1jO\xc3\xcf\x9d\xf3\xbf~\xf9\xfc\xab{\x1c5\xdc\x83R\xa5Pe\x0c\xf76\xc1\x82\x9a9\xdf\x8f\x8a\t\x01C\x84c_\xabh>\xfc\xbb\xd7\xcf\x1f\xa6\t\x8fvU\xba\xb1\xe8\xf1\xa1\xd3\xceos\x0c\xba\xe1]\x17{L\x1e\x84!H\x9c\xfb\xac?\xc4R:\xfb{\xbd\xb9\x99\xbf\tj\xfb\xbe)\xd5\x8b\x833\xe6\x7f\xd5\x84\xcbj\xc5\x04*\xb1B\x0bQ\xa7\x89\\\xff/\xa8\xa8\xfbq\xa0<\xb6e\xb5\xef\xc6Wg\x92\x97\xb4\xab\xe9e?\xd1\xfd\xde\x90\xfal\xb4o,\xa2\x92\xd5j\xfatQ\xfe\xa5\x10 \xc1Ts\x1a\x9b\xc1\x11\x10er\x89\x1a\xf3\xee#\xcd\xde\xab{kR\x0f-\xcd9g\xfeZVs\xa5\xca\x8e,\xc2\x1f\xf8\x1b!\xfb5$\xb4\x10\xcb\x84p\xfa~Ff0\x8c\x94<L\xb2\xef\x1a\xdd\xd7>\xe8\xf7\x87X\xb9b\x1b\x8d\xa6@\xefu\xa2}RZ\xcb9w\xfb\xc1\xb3\xee\x8e\xadg\xdd\x1d[\xcf\xb8\x8f\xff\xa4\xd5[\x97\xd2Q\x15\x13\x94\xa9\xe6\xfd\x92\x88t\xac\xae\xdf\xce\xe9\x9e\x01g\xd1\xe9c\xbb\x9b\xaf\x9d\xff\xad\xe78\xd3\xe9\xba\xb4\x1f\x9f(R\xb2\x86\xceK\x7f\xc1\xf1g\xd2s\xe5\xb0{\xf6\xdd\xdd\x17\xfaR\xb9O\xac\xd8F#\x80\x01\xd1\x1b9\x00\x83\xdb`\x93\xdc\xa7\xe8\xbeV\xad\x7f\xba\xcb_\xf6vV\xab)\xb4\xab"\x84\xe8%H\xbc\xee\xf9\xfd\x1aZ\xbf\x93\xab\xddi}\xf0\xf4\x89\xfdM&0\xc0\x94i\xb3\xc8v\x1d\xa6\xbf\xed\x08\xb3\x16|\x99\xf2\xaaY4\xbe\xf3BK6\xe1\x7fr\xdd7h\x1f\xe9\x82ITg\x05\x19@PvF$\xe7\x11\xea\xf0\x08\xe7\xe1]g\xcb\x88\x1b\xb4\xc9X\x00[w\x91\x1e\xcc\x12\xb2\xa5\x0f\xfc\xb9\x0c\xa5\xb3$\xe3\x0e}\x9dW\t\x86 \x93#\xb2\xfe)\x92Emx3\xed\x7fq\xf3R\x08\x17x\xd1Ic\xf7>ET\xed\xb2\x18>x\xde\xbd\xf4\xc43n"\x15wS\xa9\xa9d\x92\xedD\xa2\xb3\xaa\x9f\xff\x0e5Em(\xdc=\xeb\xcd\xf0\xe9y\xb1\xa8\xd7{\xe5\xa4\xbe\xf9Vcv\xa0?\xbeG\x05-\x8d\x1c\xff\xe6\x8a\xfa\xc5\xa1\xe93\xab\xf5\xdc\x7fZ\xeb\x81WF\x00\x14\xf6u7\x80d\xb3\xac<q\xbc+\xdc\xf0\xcfl\xc7\xac\xb9\x8f\x1f\x9a1\xef\xa1~\x80x\xef{\xbf:|\xe4\xa7k\xd5\x1b\xbc\xcfqYY\x04\xe0\xa6\xbc\xde\x86y>\xf3\xa35\x9f;R?\xbfjm\xb99\xb6C9\x0e\x80N\xb7\x88U?\xdb\xf2\xf6\xb1\xe7\x9ep\xdc\x8bkG\xaf\xb9\xa9\x06n\xc7\x02\x196\xd7\xde\xff\xe8\xfbQs\x81*s\x8a*\xf3o\xaa\xcc)\xa6\x9aF\x82\xf4\x9b9\xcb\x8e\xbe\x12\xc8\xb0\xe1\xff\x02\xe0\xd0\x81\x06\xd6\xaf{2\x01\xc3-\xaaX(f\xe4\xbfP\xe4\x07\xd6\x9eH\x8c^w\x1bw\xc1\x84\xa6iJ\xfc\x18WIim\xfef\x1bWQ\xea\x13\xf4\xc2\xf4\xf2\xe6\x81\x97\x01!\xc3\xb4\x0f\xa7\x01\x11EU\xe6\xd4\x9a\xa3\xbb\x1c)w\xfa\xf4cj3\xb8|\xec=\x92?\xbc*\xa2\xd2\xf4-O\xc2qO\xc3\x02\xe0\x13\xe8\xb1\xef\\\xff\x82M\x8ay\xd6\x8bs]"?\xba\xeeu)5W\xd01\xc7\xfbl\xebO\x95\xf7\x98w\xbe;B\x07\x86V\x7f\x8dg\x17B\xee\xc4\x0c.\x95\xe6\xcc\r$\x8e\xaey\xc1\x13\x95\xe6\xa2S_\xb1\xe9\xf2\x97X\x1fZ\x03z\x8b:\x1e\x1bW\x00i\xb8\x8b\xe7\xc0\x9d\x98)\xe0\x91\x91\xf1Ko\xb3 z\x93\xe7\xc9\xfdvR\xe7\x0c5{[\x03\xd7\x95\xabdt\x06\x1e\xc1\x8cR1&E\x01\x9e\x00\x8aK\t\xed\xdeJ\xc7\xa1<!\xf8\n\x82\xe0\x95\xce0\x7f/)\x91\x0e\n:P\x0c\x1d~\xbd\x9f\xd6\xd8\x80\xe0yy\x9f$\xb56a\xdb\x0c\x06K\xa5}\xd4\x08\xe6\xe3\x13D\xf0m\xc1\x1d\xf6\x0c\x82\xa3S\xb4\xdd[MP\xfaG\x12\xbb\x94\xd0\xe1\xd7\xff+\xa5s\x9e\x14\xbc\x9c\x8f\xa5\xa5\xb4\xed\xac\x943\x8fF\xe4\xda\x08\x97\x8a\xa1\xc7\x7f\xb0\xe3\xb2\xff\xf9m\x16\xb9\x0eP\x01\xc4\xe0$\xed\x88\xb4\x07b\xe6\r$?\x96\x0b\x1e]\xfeR:x8\x10\xc8_\xddm\xde*\x14\xcb\x1b\xa2\x02U\x0bE\x10\x94\xe1\xf9@\xe2\x19\xad:Y.-N9-\x0cI\xb4c<\xba\x05/\xe7hicHz\xaf\x15\x95\xa00\xed\x16SmF\x96\x01\xd8\x92\x05T\x04\x0f\xc1E\xf2%\x18fH\xc5&m\xa9\xe2L3\'\xb9\xca#\x13IH\x04\xcf\x82\xfc\x8cQ\x0c@\xf2I\x0b/\x95\xe1\xf8\xbc\xcd\x90\x7f`\xf0LH\xfa\x08\xd1\x97\x8f\x1f.\x81Q\xc7\xf25 E\xe7N\x91VFr\x8e\xdb\xaf\xf6\x80\xce=}\xde}\xac\xb2@\x9b\xa0d\xa8v=\x02\xad\x05\x11\x9et\x9fF1-9\xadl\xf3\xb1\xb5 B\x8f\xb0Q\xac\x0b\xbe\x04\\\x04:u9.\x91\x8b\xad\xfe\xda\xd6N].\xa3Dh\x12Z\xf7A@R\x99\x1b\x00\xc4u\xe1\xaan\x7fIht\xf3Y\xe4\xfc\x10\xf1\x1eOJ\x00h\xf16apw_\xf46\xbe\xe8\x17M\xc6\x1eAI\x0c\xd8\x0c9\x01\x82$\xa9\x03\xfc\x17Z\xbdu\x7f\xf6\x08\xca\xa8\xefM\tH2]n5\xf7euZ1\x80\x00\xc9\xae\x80$\xc7\x02CT\xf2\x1f\xb6`K\nQI\x97Hgz\\\x1a%\xcd\xfa\xcd\xcb8t\xa0\x01\xd0lX\xba\xdb\xc6\x8b\xcbj\rc\x05\xfa?\x1e\xc0\x03_\x14\xe3\t@\x00\x00\x00%tEXtdate:create\x002020-06-01T14:36:18+00:00\t\'//\x00\x00\x00%tEXtdate:modify\x002020-06-01T14:36:18+00:00xz\x97\x93\x00\x00\x00\x19tEXtSoftware\x00www.inkscape.org\x9b\xee<\x1a\x00\x00\x00\x00IEND\xaeB`\x82'
 flag_AD = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x12\x00\x00\x00\x0c\x08\x06\x00\x00\x00o\x12\xed\xbc\x00\x00\x00\x04gAMA\x00\x00\xaf\xc87\x05\x8a\xe9\x00\x00\x00\x19tEXtSoftware\x00Adobe ImageReadyq\xc9e<\x00\x00\x02\x8bIDATx\xdabdP\xf7x`\xae\xaf+\xe0hn\xc8\xc0\xc7\xc3\xc3\xf0\xff/\x13\x03\xc3/N\x86\xb4\x80\x13\x0c\x0c\xbf\x19\x18\x18\xfe1@\x00P\xf8?\x0f\x03\xc3\x93y`&\x03\x13##\xc3\xefO\x9f\x18^\xee\xdf\xcf\xf0\xe2\xe4\xc9\x0f\x00\x01\xc4\xc2\xc0\xc8"r\xf2\xeamn;3]\x06%Y\x11\x06f&\x16\xa0Al\x0c\\|j\x10\x83\xfeC\rb\x0429\x19\x18\x84\xd5 \x061\xfc\xfd\xcb\xf0\xe1\xecY\x86\x1f\'O2\xb030\xb0\x00\x04\x10\x0b\xc3\x7f\xd6?\x0c\x8cL\x0cG\xce_g`cca`ebe`\xf8\xce\xce\xe0fr\x15a\xd0\x7f\x88\x8b\x18\x80v|\xb9\na\xfe\xff\xfd\x9b\xe1\xdd\x91#`;\x80\x8e\xfe\x07\x10@@\x17q\x00E\x99\x18\x8e_x\xc0\xf0\xf4\xe5\'\x06\x16F\xa0\xea\x8f\xdc\x0cY~+\x18\xbe\x7fad\xf8\xf5\x85\x89\xe1/\xd0 6\x8e\x7f\x0c\\\xd2\xff\x19\x9e-\x828\xf0\xff\x9f?\x0c_\x1f=\x02\xfb\x1ch\x1f\x0b@\x00\x01]\xc4\x05\x14\x05jf\xfc\xc7\xf0\xe8\xf9O\x06\xc6\xff\xbf\x18\x18>\xb3001\xdfgx\xb1\x8e\x95\xe1\xddMf\x06^\xe1\xff\x0c\x7fe\xff1\xc8\xc6\xfea\xf8}\xff?\xc4\xa1\xff\xff\x83\x1d\n2\xe8/P7@\x00A\x0c\x02\xf9\x12\xec`F\xa0\x02``0\x082p\xfc\xfa\xcf\xf0\x8f\xf5?\x03\x1b\x17#\x03\'0l~\xb321\xfc\xfc\xc9\xc8\xc0\xf4\x1f\x16\xfa\x0c0o\x81\r\x02\x08 \x16p(\x821\xd0 `L\x80100X\x81>\xe6\xe6\xff\xcf\xc0\x0e\xc4\xff9\x98\x19\xfe}\xfc\xcb\xc0.\xf2\x8f\x81\x15\xa2\x1192\xc1\x11\n\x10@\xc00\xfa\x06\xd4\xfc\x07\xc8\xfb\x0f1\x04\x1aK, \x82\x1b\xe8%\x05\xa0\x80\xd0o\x86_?\xfe3\xf0}c\x04\xba\x1d\xa2\xe0/\x03".\x98\x81\x18 \x80\x80\x06}\x07\x1a\xf0\x0b\xea\x1a\x90\xf1@\xf3\xff\xb210q30hy\xfcexv\xfb?\xc3\x0f\xa0\x01\x86:\x7f\x19\x98\xc4 \x06\xb0@]\x02\xf3\x1a\xd0 F\x80\x00\x02\x8a}\x87\xfa\x90\x19\x82a\xde\x04\xea`\x06J\xc9JA=\xf1\x0b\x14\xe5\x10\x8dL\x08U`\xd7\x80h\x80\x00\x02\xba\xe8\x0f4\x9c\x80\xd4\x7f`\x08\xfc\x07\x06\xfc\x7fn\x86\xbf\x7f\xa0n\xff\x0b\xb5\x9a\x19\x14\xe5\x10!\x98k\x90\xa4\x98\x00\x02\x08\x18k\x7f\xde0\xfce\xfe\xc7\xf0\xe7?\xc4\xf0\x7f@?\xfec\xfc\xff\xf2%\xd0\xf4/\xe0T\x01\x8aiF\xb0\xd5\xc0\x00z\x05q?8\xa1Cc\x8c\xe1\x13\x03\xc3O\x80\x00\x03\x00\xe5\x7f\xcdUp(\\\xbe\x00\x00\x00\x00IEND\xaeB`\x82'
 flag_AE = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x12\x00\x00\x00\x0c\x08\x06\x00\x00\x00o\x12\xed\xbc\x00\x00\x00\x04gAMA\x00\x00\xaf\xc87\x05\x8a\xe9\x00\x00\x00\x19tEXtSoftware\x00Adobe ImageReadyq\xc9e<\x00\x00\x01\x98IDATx\xdab\x9c\xc7\xc0\xf0@\xc2\xdc\\@\xd2\xd1\x91\x81\x9f\x8f\x8fa\xb9\xff\x7f\x06\x86\x1f\x0c\x0c\x0c\xff\x18\xb0\x03F(\xc5\xc8\xc8\xf0\xe9\xd7\'\x86\xfdO\xf73\x9c|r\xf2\x03@\x00\xb1\xb030\x88\xfc8y\x92\x9b\xd5\xce\x8e\x81OI\x89AM\x80\x99\x81\xe1\x17P\xe5\x7f\x06\xdc\x00h\xd8\xdf\xff\x7f\x19\xce\xbe=\xcbp\xf2\xf3I\x06\x066\x06\x16\x80\x00b\x01\xaa\xff\x03\x92{\x7f\xe4\x08\x03?\x1b\x1b\xc3U\x03V\x06\x86\x9fx\\\x04\x05\xbf\xff\xfdf8\xf2\xea\x08\xc4\x85\xff\x18\xfe\x01\x04\x10\xcb?\xa8\x9e\xf7\xc7\x8f3p=}\xca\xb0\xc8\x81\x85\x81\xe1\x0b\x03\xc4U\xb8\x00\xc8\xf6\xff\x7f\x18\x1e}~\x04t\x1a\x90\xff\x9d\x81\x05 \x80X\xfeB\r\xfa\r\xc4\x0f\x1f=b8e\xf8\x86\x81\x18\xf0\xef\xdf?0\xfe\xf3\xe7\x0f\xc3\xdb\xb7o\x19\x01\x02\x08\xee"f\x18\r\xf4\x1e\xe3\x7fF\x86\xff\xbf\xffc\t(F\xb0\xd8\xff\xff@\x0c4\x84\xe1\xef_\x86\xff\x7f\xfe2\xfc\xf9\xf9\x8f\x11 \x80X@RLP\x0c\x02\xac@\x83\x98\x811\xf2\x8f\x15w \x81\x0c\xfa\x0b4\xe47\xd0\x1b\xff\x98\x98\x80\x963\xfd\x07\x08 \x16\x90KX\x910;33\x03\x13P\x92\x81\x05\xbf\xb7\xfe\xfee\x02\xba\x0f\xe4\xb2\xbf\x0c,@S\x00\x02\x88\x05\xa4\x1e\x84\xb9\x80X\nd\x18++\x03\x19\x80\x11 \x80@v\x83\xc3\x87\x11\xea"r\x01@\x00\xc1=\x01\n\x91?\x14\x98\x03\x10@,\xc0\xdc\xf0\x06\x187\xff\x80\xa61\xb3C\xc2\x9c\x11\x1a]\x8c\x88T\x83\x93\r\x03?\x01\x02\x0c\x00\x05+\x88\x94\x93\x89\xd7n\x00\x00\x00\x00IEND\xaeB`\x82'
 flag_AF = b'\x89PNG\r\n\x1a\n\x00\x00\x00\rIHDR\x00\x00\x00\x12\x00\x00\x00\x0c\x08\x06\x00\x00\x00o\x12\xed\xbc\x00\x00\x00\x04gAMA\x00\x00\xaf\xc87\x05\x8a\xe9\x00\x00\x00\x19tEXtSoftware\x00Adobe ImageReadyq\xc9e<\x00\x00\x02BIDATx\xdabd``x`nn.\xe0\xe8\xe8\xc8\xc0\xc7\xc7\xc7\xf0\xff\xff\x7f\x864\x06\xfc`\x96\x0b\x03\x03\xe3OF\x86O?>1\xec\x7f\xb0\x9f\xe1\xe4\x83\x93\x1f\x00\x02\x88\x05(.r\xf2\xe4In;;;\x06%%%\x06fff\x06.\x02\x06\xa9\x8900\xfc\xfd\xf9\x97\xe1\xec\x8b\xb3\x0c\'\xdf\x9cd``e`\x01\x08 \x90A\x7f@\x92G\x8e\x1ca`ccc`eeepC\xd6\xc5\xca\xce\xc0\xc0\x01\xc4\x9f?\xc1\x85\xae*10\xfc\xfe\xfa\x9b\xe1\xc8\x93#@\xa7\x01\x05\xfe1\xfc\x03\x08 \x10\xf5\x01\x88\xf9A\n\xe4\xe4\xe4\x18XXX\x18N\x808@\xcd\xff\xf5\xcd\x19\x18\xff\x02\xedabb\xf8\xff\xef\x1f\x03\xe3\x85s@k\x7f0Xt\x01\xa9w\x7f\x18\x1e\xbd}\x041\xf9\x0b\xc3w\x80\x00bA\xb6\xfc\xd1\xa3G\x0c\x8c\x8c\x8c\x0cL \xce\xb4e\x0c\x0cYQ\x0c\x0cO\x9e00\xb0\x03]$*\xca\xc0\xd0\x0f\x14+\x8cb\xb8\xff\x9a\x81\xe1\xff\xb7\xff\x10M\xff\xc1\x98\x11 \x80P\\\x04\x03\x9f\xad\xc2\x19\x18\xd3\xb2\x18\x18.?e`\xe0\xe5``\xb8s\x93\x81AL\x06\x188\xf2\x0c\xff\xe7Mb\xe0\xf5]\xc3\x00\x0e\xc8\xdf@\xfc\x17\xa4\x81\xe1\'@\x001a\x0bL6.n\x06V\xa0wX\xf7\xccc`5\x90c`\xd5\xd0a`\xdd\xb7\x80\x81\xf5?3\x03\x1b7\x1f\x03\xd8\xc9\xcc\x0c\x10\x9a\x11\x8c\xff\x03\x04\x10V\x17\x81,ax\xf3\x86\x81\xc1\xcc\x96\xe1\xbfW4\x03#\x17+0\xce\xbb\x81.\xbc\xce\xc0\xa0"\xca\xc0<\x03(\xff\r\x88\x7f@\xa3\xea3\xc3\x0f\x80\x00b\xc1\xe6"\xb03\xe3R\x18\xfe\xdf\xbd\xc6\xc0x\xf36\x03#3PYg\x19\x03\x83O\x18\x03\xc3O\xa0\x9c\x00\xd4\x10fh\x18\x01\xdd\x04\x10@,8\x13\xcb\xb6\r\x0c\x8c\xba&\xc0\xb8>\x0fT\xfc\x8f\x81A\xcf\x9c\x81\xe1\xd2ID\x003\xc0\xbd\x05\xc6\x00\x01\xc4\x08\x8e<\x06\x06nd3\xfe\x10H\x90,+\x81\xc4[ \xfe\x05\xf7\xdao\x80\x00\x02\xb9\xe8\r8I\xc1\x82\x0f\x18\xff/YX\xff\xff\x03\xd2\xc0\x94\x03\xb2\x0c\x98i\xc0\xf6\x02\x15\x00\x99\xbf\xfe32\xbc\xfd\xc7\xc0\xfc\x1e\xc8\xfe\x0b\r\xd0\xaf\x0c?\x01\x02\x0c\x00)\'\x9f0:j\x84d\x00\x00\x00\x00IEND\xaeB`\x82'
