+++ pynicotine/gtkgui/settingswindow.py
@@ -1674,6 +1674,8 @@ class IconsFrame(buildFrame):
             }
         }
 
+        self.TrayiconCheck.set_sensitive(self.frame.appindicator)
+
         self.N.set_from_pixbuf(self.frame.images["n"])
         self.Away.set_from_pixbuf(self.frame.images["away"])
         self.Online.set_from_pixbuf(self.frame.images["online"])
