+++ setup.cfg
@@ -4,6 +4,7 @@ requires =
     gobject-introspection
     gsound
     gtk3
+    libappindicator-gtk3
     libnotify
     miniupnpc >= 1.9
     pygobject3-devel
